### 2024-01-19 #6125 @matthewfeickert
@temmy222 Unfortunately, it seems that there are still problems that have been pointed out multiple times before that have not been addressed. JOSS editors have given advice on how to approach Python packaging in https://github.com/openjournals/joss-reviews/issues/4260#issuecomment-1094029318, https://github.com/openjournals/joss-reviews/issues/5203#issuecomment-1457301825, and https://github.com/openjournals/joss-reviews/issues/5442#issuecomment-1560505106. I've made it clear that you should follow the [JOSS review criteria very closely](https://joss.readthedocs.io/en/latest/review_criteria.html). The review criteria states

> Software dependencies should be clearly documented and their installation handled by an automated procedure. Where possible, software installation should be managed with a package manager. However, this criterion depends upon the maturity and availability of software packaging and distribution in the programming language being used. For example, Python packages should be `pip install`able, and have adopted [packaging conventions](https://packaging.python.org/), while Fortran submissions with a Makefile may be sufficient.

This has still not been addressed as can be seen

```console
$ docker run --rm -ti python:3.11 /bin/bash
root@8542635a357e:/# python -m venv venv && . venv/bin/activate
(venv) root@8542635a357e:/# python -m pip --quiet install --upgrade pip setuptools wheel
(venv) root@8542635a357e:# python -m pip install pytoughreact
...
Installing collected packages: pytz, tzdata, six, pyparsing, pillow, packaging, numpy, kiwisolver, fonttools, cycler, scipy, python-dateutil, contourpy, pandas, matplotlib, vtk, pytoughreact
Successfully installed contourpy-1.2.0 cycler-0.12.1 fonttools-4.47.2 kiwisolver-1.4.5 matplotlib-3.8.2 numpy-1.26.3 packaging-23.2 pandas-2.1.4 pillow-10.2.0 pyparsing-3.1.1 python-dateutil-2.8.2 pytoughreact-1.0.2 pytz-2023.3.post1 scipy-1.11.4 six-1.16.0 tzdata-2023.4 vtk-9.3.0
(venv) root@8542635a357e:# python -c 'import pytoughreact'
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/venv/lib/python3.11/site-packages/pytoughreact/__init__.py", line 9, in <module>
    from pytoughreact.writers.bio_writing import t2bio
  File "/venv/lib/python3.11/site-packages/pytoughreact/writers/bio_writing.py", line 34, in <module>
    from t2grids import t2grid, rocktype
ModuleNotFoundError: No module named 't2grids'
(venv) root@8542635a357e:#
```

While you do mention [in the project README](https://github.com/temmy222/PyTOUGHREACT/blob/a25d2435d35c1226113db0775dc0696ea4b2a62b/README.md?plain=1#L34)

> Because pytoughreact requires PyTOUGH and PyTOUGH is not uploaded to PyPI, it is required to download the zip folder of PyTOUGH from the GitHub repository https://github.com/acroucher/PyTOUGH. Unzip the folder and place in your current working directory. Change directory into the PyTOUGH folder and run python setup.py install or pip install on the command line. With PyTOUGH installed, PyTOUGHREACT is ready to be used as a package.

this isn't an acceptable method given what we've already established. While it is not great, but perhaps allowable, to use an unpacked dependency like PyTOUGH please include it as a requirement in your [`setup.cfg` `install_requires` metadata](https://github.com/temmy222/PyTOUGHREACT/blob/a25d2435d35c1226113db0775dc0696ea4b2a62b/setup.cfg#L22C17-L22C18) like the following

```txt
install_requires =
    numpy>=1,
    scipy>=1,
    vtk>=1,
    matplotlib>=3,
    pandas>=1
    PyTOUGH @ git+https://github.com/acroucher/PyTOUGH.git@v.1.5.7
```

You'll have to pin at a version rather than deal with whatever is coming down the pipe on the `HEAD` of the default branch. I'd encourage you to ask the authors to package the software though (there aren't any such requests on their Issue tracker: https://github.com/acroucher/PyTOUGH/issues) and maybe adopt standard semantic versioning (though their project isn't your responsibility, to be clear).

```console
$ docker run --rm -ti python:3.11 /bin/bash
root@14297e51f124:/# python -m venv venv && . venv/bin/activate
(venv) root@14297e51f124:/# python -m pip --quiet install --upgrade pip setuptools wheel
(venv) root@14297e51f124:/# git clone https://github.com/temmy222/PyTOUGHREACT.git
Cloning into 'PyTOUGHREACT'...
remote: Enumerating objects: 2256, done.
remote: Total 2256 (delta 555), reused 798 (delta 447), pack-reused 1324
Resolving deltas: 100% (1322/1322), done.
(venv) root@14297e51f124:/# cd PyTOUGHREACT
(venv) root@14297e51f124:/PyTOUGHREACT# git reset --hard a25d2435d35c1226113db0775dc0696ea4b2a62b
HEAD is now at a25d243 requirements file
(venv) root@14297e51f124:/PyTOUGHREACT# apt update && apt install -y vim
...
(venv) root@14297e51f124:/PyTOUGHREACT# vi setup.cfg
(venv) root@14297e51f124:/PyTOUGHREACT# git diff
diff --git a/setup.cfg b/setup.cfg
index ec282dc..2baabb5 100644
--- a/setup.cfg
+++ b/setup.cfg
@@ -25,6 +25,7 @@ install_requires =
     vtk>=1,
     matplotlib>=3,
     pandas>=1
+    PyTOUGH @ git+https://github.com/acroucher/PyTOUGH.git@v.1.5.7
 python_requires = >=3.7
 package_dir =
     =src

(venv) root@14297e51f124:/PyTOUGHREACT# python -m pip install .
...
Successfully built pytoughreact PyTOUGH
Installing collected packages: pytz, PyTOUGH, tzdata, six, pyparsing, pillow, packaging, numpy, kiwisolver, fonttools, cycler, scipy, python-dateutil, contourpy, pandas, matplotlib, vtk, pytoughreact
Successfully installed PyTOUGH-1.5.7 contourpy-1.2.0 cycler-0.12.1 fonttools-4.47.2 kiwisolver-1.4.5 matplotlib-3.8.2 numpy-1.26.3 packaging-23.2 pandas-2.1.4 pillow-10.2.0 pyparsing-3.1.1 python-dateutil-2.8.2 pytoughreact-0.0.15 pytz-2023.3.post1 scipy-1.11.4 six-1.16.0 tzdata-2023.4 vtk-9.3.0
(venv) root@14297e51f124:/PyTOUGHREACT# python -c 'import pytoughreact; print(pytoughreact)'
<module 'pytoughreact' from '/venv/lib/python3.11/site-packages/pytoughreact/__init__.py'>
(venv) root@14297e51f124:/PyTOUGHREACT#
```

As I've given you instructions on what needs to be done, please make these corrections, make a new release on GitHub and PyPI, **verify that all other previous requests have been fully addressed** and then please notify us on this Issue within 1 week (2024-01-26). Otherwise I am afraid we will have to reject this for a 4th time.
