### 2024-02-04 0437d42
update setup.cfg
