+++ setup.cfg
@@ -3,7 +3,9 @@ name = pytoughreact
 description = Python Library for automating reaction simulations using TOUGHREACT, TMVOC and TMVOC-BIO
 author = Temitope Ajayi
 author_email = ajayi_temmy@yahoo.com
-version = 0.0.15
+long_description = file: documentation/technical.rst
+version = 1.0.2
+url= https://github.com/temmy222/PyTOUGHREACT/tree/master
 license = MIT
 license_files = LICENSE 
 platforms = unix, linux, osx, cygwin, win32
@@ -20,11 +22,12 @@ classfiers =
 packages =
     find:
 install_requires =
-    numpy>=1, 
-    scipy>=1, 
-    vtk>=1, 
-    matplotlib>=3,
+    numpy>=1
+    scipy>=1
+    vtk>=1
+    matplotlib>=3
     pandas>=1
+    PyTOUGH @ git+https://github.com/acroucher/PyTOUGH.git@v.1.5.7
 python_requires = >=3.7
 package_dir =
     =src
