### 2024-02-05 #6125 @temmy222
Hi @matthewfeickert  I tried to upload the package to PyPI by adding PyTOUGH @ git+https://github.com/acroucher/PyTOUGH.git@v.1.5.7 to the setup.cfg folder but it seems this isnt allowed by PyPI for uploads to PyPI. Some links online also seem to suggest this process isnt allowed
https://stackoverflow.com/questions/54887301/how-can-i-use-git-repos-as-dependencies-for-my-pypi-package
https://github.com/pypi/warehouse/issues/7136

I couldnt find any other way to bypass this rule. Is there another solution that may be of help?

Thank you
