### 2024-02-05 #6125 @kyleniemeyer
Hello @temmy222. Can you try the directions given here: https://chrisholdgraf.com/blog/2022/install-github-from-pyproject/

Also, since you now have a `pyproject.toml` file (the current recommended Python packaging approach), you should probably remove `setup.py`. I would also recommend moving all your dependency specifications into `pyproject.toml` and eliminating both `requirements.txt` and `setup.cfg` - all of that should be able to go inside `pyproject.toml`, and it would clean things up significantly to have just one file.

I also see a bunch of misc files (like `.png` files) in your repository. Can you clean that up? You should really just have the core package files, unless necessary for documentation (which would go in a subfolder).
