### 2024-02-06 #6125 @matthewfeickert
> I tried to upload the package to PyPI by adding `PyTOUGH @ git+https://github.com/acroucher/PyTOUGH.git@v.1.5.7` to the `setup.cfg` file but it seems this isn't allowed by PyPI for uploads to PyPI.

Ah. So @temmy222 you're running into https://github.com/pypi/warehouse/issues/7136, which is a reflection of the PyPA's [stance on direct references in public indexes](https://packaging.python.org/en/latest/specifications/version-specifiers/#direct-references)

> Public index servers SHOULD NOT allow the use of direct references in uploaded distributions. Direct references are intended as a tool for software integrators rather than publishers.

, correct? I wasn't aware of this when I had suggested this to you, as I'm only used to using direct references in paritcular dev testing workflows or applications &mdash; not package deployment.

However, looking at the definition of a direct reference

> A direct reference consists of the specifier `@` and an explicit URL

you might be able to use

```
git+https://github.com/acroucher/PyTOUGH.git
```

as this doesn't use the `@` specifier. My recommendation is to try it, though I will consult packaging experts who might know as well.

Please do also consider @kyleniemeyer's suggestions on modernizing your packaging specification (I would recommend reading the [Scientific Python organization's development guide section on simple packaging](https://learn.scientific-python.org/development/guides/packaging-simple/) for this) and also address his points on the repository organization. Please also remove the multiple files that were added in https://github.com/temmy222/PyTOUGHREACT/commit/cfe9ef4e22a55a43beddd442c34a5b5374241c7c that are duplicates with a commit SHA in the name.
