### 2024-02-20 #6125 @matthewfeickert
@temmy222 As I asked the `PyTOUGH` maintainers to release the project on PyPI (c.f. https://github.com/acroucher/PyTOUGH/issues/55) they did, very graciously, with `v1.6.0`.

https://pypi.org/project/PyTOUGH/

I asked you back in January to do this in https://github.com/openjournals/joss-reviews/issues/6125#issuecomment-1901093941

> I'd encourage you to ask the authors to package the software though (there aren't any such requests on their Issue tracker: https://github.com/acroucher/PyTOUGH/issues) and maybe adopt standard semantic versioning (though their project isn't your responsibility, to be clear).

Is there a reason that you didn't?

Please update `PyToughReact`'s `setup.cfg` to use the version of `PyTOUGH` on PyPI

```
PyTOUGH>=1.6.0
```

and make a new release of `PyToughReact` with this change on GitHub and PyPI. Please do this _before_ Monday 2024-02-26 and let us know here once you've done so. :+1:
