### 2024-02-20 #6125 @acroucher
hi @temmy222 , just to let you know I have just released PyTOUGH 1.6.0 which is installable from PyPI using `pip install PyTOUGH`. Hopefully this will help resolve your own packaging problems. Apologies for the trouble...
