+++ setup.cfg
@@ -4,7 +4,7 @@ description = Python Library for automating reaction simulations using TOUGHREAC
 author = Temitope Ajayi
 author_email = ajayi_temmy@yahoo.com
 long_description = file: documentation/technical.rst
-version = 1.0.2
+version = 1.0.3
 url= https://github.com/temmy222/PyTOUGHREACT/tree/master
 license = MIT
 license_files = LICENSE 
@@ -27,7 +27,9 @@ install_requires =
     vtk>=1
     matplotlib>=3
     pandas>=1
-    PyTOUGH @ git+https://github.com/acroucher/PyTOUGH.git@v.1.5.7
+    PYTOUGH @ git+https://git@github.com/acroucher/PYTOUGH.git
+
+
 python_requires = >=3.7
 package_dir =
     =src
