### 2024-02-26 2aa8fac
update pyproject.toml and tox
