+++ tox.ini
@@ -22,7 +22,7 @@ deps =
     sphinx_rtd_theme
     pytest
     pytest-cov
-    PYTOUGH @ git+https://git@github.com/acroucher/PYTOUGH.git
+    PyTOUGH
 allowlist_externals=*
 commands =
     pytest --verbose
