+++ setup.cfg
@@ -1,56 +1,0 @@
-[metadata]
-name = pytoughreact
-description = Python Library for automating reaction simulations using TOUGHREACT, TMVOC and TMVOC-BIO
-author = Temitope Ajayi
-author_email = ajayi_temmy@yahoo.com
-long_description = file: documentation/technical.rst
-version = 1.0.3
-url= https://github.com/temmy222/PyTOUGHREACT/tree/master
-license = MIT
-license_files = LICENSE 
-platforms = unix, linux, osx, cygwin, win32
-classfiers = 
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.7
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    Programming Language :: Python :: 3.11
-
-[options]
-packages =
-    find:
-install_requires =
-    numpy>=1
-    scipy>=1
-    vtk>=1
-    matplotlib>=3
-    pandas>=1
-    PYTOUGH @ git+https://git@github.com/acroucher/PYTOUGH.git
-
-
-python_requires = >=3.7
-package_dir =
-    =src
-zip_safe = no
-[options.packages.find]
-where=src
-
-[options.extras_require]
-testing =
-    pytest>=6.0
-    pytest-cov>=2.0
-    mypy>=0.910
-    flake8>=3.9
-    tox>=3.24
-
-[options.package_data]
-pytoughreact = py.typed
-
-[flake8]
-max-line-length = 160
-per-file-ignores =
-    */__init__.py: F401, E402
-    */results/result_tough_3.py: E731
-    */results/result_single.py: E721  
