### 2024-02-27 #6125 @matthewfeickert
> Please let me know if I am missing any requested change.

@temmy222 as both @kyleniemeyer and I already noted, your repository has multiple things that appear to be miscellaneous files or junk files that have no use that should not be in the repostiory

> I also see a bunch of misc files (like `.png` files) in your repository. Can you clean that up? You should really just have the core package files, unless necessary for documentation (which would go in a subfolder).

> Please also remove the multiple files that were added in [temmy222/PyTOUGHREACT@cfe9ef4](https://github.com/temmy222/PyTOUGHREACT/commit/cfe9ef4e22a55a43beddd442c34a5b5374241c7c) that are duplicates with a commit SHA in the name.
