+++ pyproject.toml
@@ -28,7 +28,7 @@ classifiers = [
   "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
-  "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
+  "License :: OSI Approved :: MIT License",
   "Operating System :: OS Independent"
 ]
 
