### 2024-03-13 #6125 @matthewfeickert
@temmy222 As I'm continuing to look for reviewers, please also take time to check to make sure that the things we've already asked you to clean up multiple times are in fact done. When you are submitting a project to be reviewed it is assumed that you have already done the due diligence to have it be ready.

Why is https://github.com/temmy222/PyTOUGHREACT/blob/4ef403e2993424c7b5e97dee9ef6522a272e8617/MESH still at the top level of the repository and why does https://github.com/temmy222/PyTOUGHREACT/blob/4ef403e2993424c7b5e97dee9ef6522a272e8617/README.md#installation still refer to local installs of `PyTOUGH`?
