### 2024-03-14 #6486 @editorialbot
[REVIEW]: PyToughReact – A Python Package for automating reactive transport and biodegradation simulations.
**Submitting author:** @temmy222 (<a href="http://orcid.org/0000-0002-0782-7460">Temitope Ajayi</a>)
**Repository:** https://github.com/temmy222/PyTOUGHREACT
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
