### 2024-03-15 temmy222/PyTOUGHREACT#6 @frank1010111
JOSS review: Installation instructions
Hello,

This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6486).

The installation instructions are different for the [README](https://github.com/temmy222/PyTOUGHREACT/blob/master/README.md#installation) versus the [documentation](https://pytoughreact.readthedocs.io/en/latest/installation.html). Specifically, the README says that pyTOUGH is not available on PyPI and must be installed separately. The documentation says that this is not necessary.

From looking at the [pyproject.toml](https://github.com/temmy222/PyTOUGHREACT/blob/master/pyproject.toml#L20), I'm leaning towards the documentation being right and the README needing correction. Either way, clarification would be great.
