### 2024-03-15 #6486 @frank1010111
Hi @temmy222

I have a question on authorship: I see that you're responsible for all the code commits and are listed as the sole author of the python package, but you're sharing authorship with Ipsita Gupta on the JOSS paper. I know that [the official JOSS guidelines on authorship](https://joss.readthedocs.io/en/latest/submitting.html#authorship) allow for non-code contributions to count, but could you confirm Dr. Gupta's contribution to this publication?

Thank you
