### 2024-03-15 temmy222/PyTOUGHREACT#7 @frank1010111
JOSS review: documentation improvements
This is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/6486).

I have identified a few issues with documentation. Apologies for the long list.
- In [Why PyTOUGHREACT](https://pytoughreact.readthedocs.io/en/latest/technical.html#why-pytoughreact), the link to PyTOUGH is dead: <https://github.com/acroucher/PyTOUGH/blob/master/doc/PyTOUGH-guide.pdf>
- There are no docstrings for
  - [Chemical Kinetic Properties](https://pytoughreact.readthedocs.io/en/latest/chemical.html#chemical-kinetic-properties),
  - [Chemical Mineral composition](https://pytoughreact.readthedocs.io/en/latest/chemical.html#chemical-mineral-composition),
  - [Chemical Mineral Zone](https://pytoughreact.readthedocs.io/en/latest/chemical.html#chemical-mineral-zone),
  - [Chemical Permeability Porosity Zone](https://pytoughreact.readthedocs.io/en/latest/chemical.html#chemical-permeability-porosity-zone),
  - any [exceptions](https://pytoughreact.readthedocs.io/en/latest/exceptions.html#exceptions-class)
  - The [Grid](https://pytoughreact.readthedocs.io/en/latest/pytoughreact.model.html#pytoughreact.model.grid.Grid) object, which seems important
  - [Relative permeability](https://pytoughreact.readthedocs.io/en/latest/exceptions.html#exceptions-class)
  - [Results for Experimental Data](https://pytoughreact.readthedocs.io/en/latest/results.html#results-for-experimental-data)
  - [Results for t2result](https://pytoughreact.readthedocs.io/en/latest/results.html#results-for-t2result)
  - [Utilities](https://pytoughreact.readthedocs.io/en/latest/utilities.html#utilities-class)
  - [Wrappers](https://pytoughreact.readthedocs.io/en/latest/wrapper.html)
- Docstring confusion
  - Most of the docstrings provide insufficient information to use the method, class, or function. If I might suggest putting yourself in the mindset of a new PhD student, excited to be using your tool to run Tough React but overwhelmed by all the options you've given them and trying to understand, say, the way to set biomass compositions.
  - The docstrings for the [Constants Class](https://pytoughreact.readthedocs.io/en/latest/constants.html) are just copies of the MIT license
  - Several of the parameters in [React Writing](https://pytoughreact.readthedocs.io/en/latest/writers.html#react-writing) are missing explanations
  - Parameters are also missing from [Biomass Composition](https://pytoughreact.readthedocs.io/en/latest/chemical.html#biomass-composition). I'm not sure why the docstring says "rock type".
  - It is unclear whether [duplicate_index](https://pytoughreact.readthedocs.io/en/latest/results.html#pytoughreact.results.result_tough_3.ResultTough3.duplicate_index) ascertains whether an index is duplicated or creates a duplicate index
  - It is unclear what type of file is read by [read_file](https://pytoughreact.readthedocs.io/en/latest/results.html#pytoughreact.results.result_tough_3.ResultTough3.read_file)
- In [Contributions](https://pytoughreact.readthedocs.io/en/latest/contributing.html), how does one run `flake8`? Are any flake8 plugins used? Is it run under default settings?
- Also in Contributions, it would be nice to link to the [Developer Notes](https://pytoughreact.readthedocs.io/en/latest/developer.html)
- In [Dependencies](https://pytoughreact.readthedocs.io/en/latest/dependencies.html), the list of dependencies is different than in the [pyproject.toml](https://github.com/temmy222/PyTOUGHREACT/blob/4ef403e2993424c7b5e97dee9ef6522a272e8617/pyproject.toml#L14-L20)
- In the [README badge](https://github.com/temmy222/PyTOUGHREACT/blob/4ef403e2993424c7b5e97dee9ef6522a272e8617/README.md?plain=1#L205) and PyPI site, the license is LGPL
![image](https://github.com/temmy222/PyTOUGHREACT/assets/4694479/3ce0bd8d-1815-420e-bd86-844fb023cb04)
In other places, it's listed as MIT. I think the LICENSE file that says MIT has primacy, but it would be nice to be sure
- The paper is roughly 2,000 words, but [JOSS suggests](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) 250-1000 words.

I haven't gotten to the tutorial yet, so I don't have any comments or suggestions there right now.
