### 2024-03-16 temmy222/PyTOUGHREACT#7 @frank1010111
Another suggestion on [Contributions](https://pytoughreact.readthedocs.io/en/latest/contributing.html), could you include information on how to install `flake8` and `pytest`?
