### 2024-03-16 temmy222/PyTOUGHREACT#8 @frank1010111
JOSS review: testing
Hello again, this is an issue related to [the JOSS review](https://github.com/openjournals/joss-reviews/issues/6486).

Neither [Contributions](https://pytoughreact.readthedocs.io/en/latest/contributing.html) nor the readme give sufficient information to run the associated tests.

Could you include information on how to install `flake8` and `pytest`?

Installing pytest with pip and running it leads to this message:
```sh
$ pytest
ERROR: usage: pytest [options] [file_or_dir] [file_or_dir] [...]
pytest: error: unrecognized arguments: --cov=pytoughreact
```

That seems to indicate that pytest-cov is another testing dependency. Looking at the [tox.ini](https://github.com/temmy222/PyTOUGHREACT/blob/master/tox.ini#L24) seems to confirm that. This also leads me to wonder: Do you recommend using `tox` to run the unit tests?

Looking at the tests themselves, there don't appear to be any that run the simulations or check simulation outputs:
```sh
$ rg \\.run test
test/test_pytoughreact_bio.py
121:        # bio.run(simulator='tmvoc', runlocation='')

test/test_pytoughreact_react.py
339:        # react.run(writeSolute, simulator='treacteos1.exe')
```
