### 2024-03-23 d57a245
fix installation instructions in README
