### 2024-03-29 temmy222/PyTOUGHREACT#7 @frank1010111
Thank you,

It would also be very helpful for the docs and readme to talk about the TOUGHREACT binaries. I'm guessing you have to at least get the [TOUGHREACT binary](https://tough.lbl.gov/licensing-download/toughreact-licensing-download/). But also maybe [tough2](https://tough.lbl.gov/licensing-download/itough2-price-sheet/) for tmvoc?
