### 2024-03-29 temmy222/PyTOUGHREACT#7 @frank1010111
Okay, I'm getting to the tutorials now. My first question concerns these lines:

https://github.com/temmy222/PyTOUGHREACT/blob/d57a245128f79d3a8a7fdcd866eca1b44f36f146/documentation/tutorial.rst?plain=1#L54-L76

These properties t2react of don't appear to be [documented](https://pytoughreact.readthedocs.io/en/latest/writers.html#pytoughreact.writers.react_writing.t2react). Are they required? What  is the react parameter dictionary? What keys are allowed and what do they mean?
