### 2024-03-29 temmy222/PyTOUGHREACT#7 @frank1010111
For the grid class part:

https://github.com/temmy222/PyTOUGHREACT/blob/d57a245128f79d3a8a7fdcd866eca1b44f36f146/documentation/tutorial.rst?plain=1#L85-L91

What do the values for `rocktype` mean? `rocktype` is documented nowhere I can find, and it appears to be pulled from PyTOUGH, which has [limited information](https://github.com/acroucher/PyTOUGH/blob/master/t2grids.py#L20-L22) on the properties. Things like the units of each measure are missing even from their code.

Then you delete a rock type and add in sand. What is the default rock type? Does it always get deleted?

What are the differences between this grid that you call t2reactgrid, and the Grid object? It took me a while to figure out that they were distinct entities.
