+++ tox.ini
@@ -9,7 +9,8 @@ python =
     3.10: py310
     3.11: py311
 
-[testenv] 
+[testenv]
+description = run unit tests
 setenv =
     PYTHONPATH = {toxinidir}
 deps =
@@ -23,12 +24,14 @@ deps =
     pytest
     pytest-cov
     PyTOUGH
+    flake8
 allowlist_externals=*
 commands =
     pytest --verbose
 
 
 [testenv:docs]
+description = build documentation
 changedir = documentation
 deps =
     sphinx
@@ -37,10 +40,12 @@ commands =
 basepython = python3.10
 
 
-; [testenv::flake8]
-; basepython = python3.7
-; deps = flake8
-; commands = flake8 src
+[testenv:flake8]
+basepython = python3.9
+deps = 
+    flake8
+commands = 
+    flake8 src tests
 
 ; [testenv:mypy]
 ; basepython = python3.7
