+++ src/pytoughreact/plotting/plot_multiple_tough_routine.py
@@ -59,8 +59,8 @@ class PlotMultiTough(object):
             restart file location
         experiment : str
             experimental file location
-        
-        
+
+
         Returns
         --------
 
@@ -76,16 +76,20 @@ class PlotMultiTough(object):
 
     def read_file(self):
         os.chdir(self.filelocations)
-        if self.simulatortype.lower() == gc.TMVOC or self.simulatortype.lower() == gc.TOUGH3:
-            fileReader = ResultTough3(self.simulatortype, self.filelocations, self.filetitles,
+        if (self.simulatortype.lower() == gc.TMVOC or
+                self.simulatortype.lower() == gc.TOUGH3):
+            fileReader = ResultTough3(self.simulatortype, self.filelocations,
+                                      self.filetitles,
                                       generation=self.generation)
         else:
-            fileReader = ResultReact(self.simulatortype, self.filelocations, self.filetitles)
+            fileReader = ResultReact(self.simulatortype, self.filelocations,
+                                     self.filetitles)
         return fileReader
 
     def read_file_multi(self, file, filetitle):
         os.chdir(file)
-        if self.simulatortype.lower() == gc.TMVOC or self.simulatortype.lower() == gc.TOUGH3:
+        if (self.simulatortype.lower() == gc.TMVOC or
+                self.simulatortype.lower() == gc.TOUGH3):
             fileReader = ResultTough3(self.simulatortype, file, filetitle)
         else:
             fileReader = ResultReact(self.simulatortype, file, filetitle)
@@ -101,10 +105,13 @@ class PlotMultiTough(object):
         locations = self.getRestartLocations()
         final_time = []
         for i in range(0, len(locations)):
-            if self.simulatortype.lower() == gc.TMVOC or self.simulatortype.lower() == gc.TOUGH3:
-                fileReader = ResultTough3(self.simulatortype, locations[i], self.filetitle)
+            if (self.simulatortype.lower() == gc.TMVOC or
+                    self.simulatortype.lower() == gc.TOUGH3):
+                fileReader = ResultTough3(self.simulatortype, locations[i],
+                                          self.filetitle)
             else:
-                fileReader = ResultReact(self.simulatortype, locations[i], self.filetitles)
+                fileReader = ResultReact(self.simulatortype, locations[i],
+                                         self.filetitles)
             if i == 0:
                 time_year = fileReader.convert_times(format_of_date)
                 final_time.append(time_year)
@@ -119,22 +126,29 @@ class PlotMultiTough(object):
         locations = self.getRestartLocations()
         final_result = []
         for i in range(0, len(locations)):
-            if self.simulatortype.lower() == gc.TMVOC or self.simulatortype.lower() == gc.TOUGH3:
-                fileReader = ResultTough3(self.simulatortype, locations[i], self.filetitles)
+            if (self.simulatortype.lower() == gc.TMVOC or
+                    self.simulatortype.lower() == gc.TOUGH3):
+                fileReader = ResultTough3(self.simulatortype, locations[i],
+                                          self.filetitles)
             else:
-                fileReader = ResultReact(self.simulatortype, locations[i], self.filetitles)
+                fileReader = ResultReact(self.simulatortype, locations[i],
+                                         self.filetitles)
             if i == 0:
-                result_array = fileReader.get_timeseries_data(param, gridblocknumber)
+                result_array = fileReader.get_timeseries_data(param,
+                                                              gridblocknumber)
                 final_result.append(result_array)
             else:
-                result_array = fileReader.get_timeseries_data(param, gridblocknumber)
+                result_array = fileReader.get_timeseries_data(param,
+                                                              gridblocknumber)
[…]
