+++ src/pytoughreact/model/relative_permeability.py
@@ -36,8 +36,11 @@ class RelativePermeability(object):
         self.validate_rel_perm()
 
     def validate_rel_perm(self):
-        rel_perm_aggregates = [rel_perm_constants.REL_PERM_COREY, rel_perm_constants.REL_PERM_EXPONENTIAL, rel_perm_constants.REL_PERM_FAT_KLIKOFF,
-                               rel_perm_constants.REL_PERM_GRANT, rel_perm_constants.REL_PERM_LINEAR, rel_perm_constants.REL_PERM_MOBILE_PHASES,
+        rel_perm_aggregates = [rel_perm_constants.REL_PERM_COREY, rel_perm_constants.REL_PERM_EXPONENTIAL,
+                               rel_perm_constants.REL_PERM_FAT_KLIKOFF,
+                               rel_perm_constants.REL_PERM_GRANT,
+                               rel_perm_constants.REL_PERM_LINEAR,
+                               rel_perm_constants.REL_PERM_MOBILE_PHASES,
                                rel_perm_constants.REL_PERM_VAN_GENUCHTEN, rel_perm_constants.REL_PERM_VERMA]
         if self.type_rel.upper() not in rel_perm_aggregates:
             raise RelativePermeabilityTypeError(self.type_rel, rel_perm_aggregates)
@@ -68,8 +71,10 @@ class RelativePermeability(object):
             raise RestrictionError([self.parameters[0], 1], 'greater')
 
     def rel_perm_converter(self):
-        mapping_rel_perm = {rel_perm_constants.REL_PERM_LINEAR: 1, rel_perm_constants.REL_PERM_EXPONENTIAL: 2, rel_perm_constants.REL_PERM_COREY: 3,
-                            rel_perm_constants.REL_PERM_GRANT: 4, rel_perm_constants.REL_PERM_MOBILE_PHASES: 5, rel_perm_constants.REL_PERM_FAT_KLIKOFF: 6,
+        mapping_rel_perm = {rel_perm_constants.REL_PERM_LINEAR: 1, rel_perm_constants.REL_PERM_EXPONENTIAL: 2,
+                            rel_perm_constants.REL_PERM_COREY: 3,
+                            rel_perm_constants.REL_PERM_GRANT: 4, rel_perm_constants.REL_PERM_MOBILE_PHASES: 5,
+                            rel_perm_constants.REL_PERM_FAT_KLIKOFF: 6,
                             rel_perm_constants.REL_PERM_VAN_GENUCHTEN: 7, rel_perm_constants.REL_PERM_VERMA: 8}
         rel_perm = {'type': mapping_rel_perm[self.type_rel.upper()], 'parameters': self.parameters}
         return rel_perm
