+++ tox.ini
@@ -1,3 +1,5 @@
+; run individual tests with tox -e for xample tox -e flake8
+
 [tox]
 minversion = 3.8.0
 envlist = py39, py310, py311
@@ -41,11 +43,18 @@ basepython = python3.10
 
 
 [testenv:flake8]
-basepython = python3.9
+description = run flake8
+skip_install = true
 deps = 
     flake8
 commands = 
-    flake8 src tests
+    flake8 src
+
+
+[flake8]
+ignore = E721, W504, F401, E402, E731
+extend-ignore = W504, W503
+max-line-length=120
 
 ; [testenv:mypy]
 ; basepython = python3.7
