+++ .github/workflows/tests.yml
@@ -22,5 +22,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install tox tox-gh-actions
-    - name: Test with tox
+    - name: Unit test with tox
       run: tox
+    - name: Run Flake 8 
+      run: tox -e flake8
