### 2024-03-29 8eac5a1
modify README for pytests
