+++ README.md
@@ -189,25 +189,31 @@ pip install -e .
 ```
 3. Modify the code 
 4. Tests are conducted with pytest and coverage reports are performed using pytest-cov. Install pytest and pytest-cov using the commands below
-   ```python
+   
+```python
 pip install pytest
 ```
+
 ```python
 pip install pytest-cov
 ```
 5. Run tests:  Run the below command from the root folder to run the tests
+   
 ```python
 pytest
 ```
+
 6. Flake8 is also used to ensure code readability. Install flake8 using 
+   
 ```python
 pip install flake
 ```
 and run flake8 using
+
 ```python
 flake8 src
 ```
-7. Make a pull request after passing all tests
+8. Make a pull request after passing all tests
 
 ## Documentation
 Documentation can be found here https://pytoughreact.readthedocs.io/en/latest/ 
