### 2024-03-29 temmy222/PyTOUGHREACT#8 @temmy222
Hi @frank1010111 . The README has been updated to show how to install flake8 and pytests in order to run the test. The idea of the tests is not to run the simulations or check the outputs but to check that the software writes to the required file and can read all values from the files into the classes in the software.
