### 2024-04-06 fa9b4a9
modify bio process doc strings
