### 2024-04-06 2cfc682
modify chemical.rst for init methods
