### 2024-04-06 6416ed8
modify bio doctrings
