+++ src/pytoughreact/chemical/bio_process_description.py
@@ -92,28 +92,27 @@ class BIODG(object):
 
 
 class Process(object):
-    """Process specification
-
-    Initialization of Parameters
-
-    Parameters
-    -----------
-    biomass :  Biomass
-        This should be a Biomass class with all properties of the biomass
-    numberOfComponents : int
-        Number of Components present in the simulation
-    mumax: float
-        Maximum specific substrate degradation rate
-
-    Returns
-    --------
-    num_of_competiting : int
-        Number of Competiting species
-    """
+    """Process specification"""
 
     def __init__(self, biomass, numberOfComponents, mumax, yield_mass, enthalpy, totalComp=0,
                  NumOfHaldane=0, NumOfNonCompetiting=0,
                  NumOfCompetiting=0, componentParams=None, gasParams=None, waterParams=None):
+        """Initialization of Parameters
+
+        Parameters
+        -----------
+        biomass :  Biomass
+            This should be a Biomass class with all properties of the biomass
+        numberOfComponents : int
+            Number of Components present in the simulation
+        mumax: float
+            Maximum specific substrate degradation rate
+
+        Returns
+        --------
+        num_of_competiting : int
+            Number of Competiting species
+        """
         self.enthalpy = enthalpy
         self.numberOfComponents = numberOfComponents
         self.yield_mass = yield_mass
