### 2024-04-06 295a113
test auto method
