+++ documentation/chemical.rst
@@ -6,9 +6,12 @@ Biodegradation Processes
 .. autoclass:: pytoughreact.chemical.bio_process_description.BIODG
     :members:
     :special-members:
+    .. automethod:: __init__
 .. autoclass:: pytoughreact.chemical.bio_process_description.Process
     :members:
     :special-members:
+    .. automethod:: __init__
+    
 ++++++++++++++++++++++
 Biomass Composition
 ++++++++++++++++++++++
