### 2024-04-06 cc818b3
test conf.py changes for documenting init method
