+++ src/pytoughreact/chemical/kinetic_properties.py
@@ -104,6 +104,48 @@ class Dissolution(Kinetic):
 class Precipitation(Kinetic):
     def __init__(self, rateConstant, ratepH, exponentN, exponentTheta, activationEnergy, coefA, coefB, coefC,
                  initVolumeFraction, prepLawIndex, logQKgap, tempGap1, tempGap2):
+        """Initialization of Parameters
+
+        Parameters
+        -----------
+        rateConstant :  float
+            Rate constant (in mol/m2/sec) at 25°C
+        ratepH : int
+            Flag for rate constant dependence on pH
+        exponentN : float
+            exponent eta in rate equation
+        exponentTheta : float
+            exponent theta in rate equation
+        activationEnergy : float
+            Activation energy in kJ/mol
+        coefA : float
+            Coefficient that shows the rate constant dependence on temperature. Defaulted to zero
+            unless a different form of rate constant dependence with temperature is desired
+        coefB : float
+            Coefficient that shows the rate constant dependence on temperature. Defaulted to zero
+            unless a different form of rate constant dependence with temperature is desire
+        coefC : float
+            Coefficient that shows the rate constant dependence on temperature. Defaulted to zero
+            unless a different form of rate constant dependence with temperature is desire
+        initVolumeFraction : float
+            The initial volume fraction (Vmineral/Vsolid) to be assumed for calculating initial effective surface
+            are a if the mineral is not present at the start of a simulation but precipitates as a new reaction
+            product
+        prepLawIndex : int
+            Precipitation law index (See user guide for more)
+        logQKgap : float
+            Log (Q/K) gap (supersaturation window). A zero value represents no gap (See user guide for more).
+        tempGap1 : float
+            Temperature (in °C) at which to begin reducing gap
+        tempGap2 : float
+            Temperature (in °C) endpoint at which the gap has diminished to nearly zero (1% of original
+            value). The gap decreases exponentially from the first (tempGap1) to the second (tempGap2)
+            temperature, and tempGap2 must always be greater than tempGap1
+
+        Returns
+        --------
+
+        """
         super(Precipitation, self).__init__(rateConstant, ratepH, exponentN, exponentTheta, activationEnergy, coefA,
                                             coefB, coefC)
         self.tempGap2 = tempGap2
@@ -116,6 +158,23 @@ class Precipitation(Kinetic):
 
 class Equilibrium(object):
     def __init__(self, logQK, tempGap1, tempGap2):
+        """Initialization of Parameters
+
+        Parameters
+        -----------
+        logQK : float
+            Log (Q/K) gap (supersaturation window). A zero value represents no gap (See user guide for more).
+        tempGap1 : float
+            Temperature (in °C) at which to begin reducing gap
+        tempGap2 : float
+            Temperature (in °C) endpoint at which the gap has diminished to nearly zero (1% of original
+            value). The gap decreases exponentially from the first (tempGap1) to the second (tempGap2)
+            temperature, and tempGap2 must always be greater than tempGap1
+
+        Returns
+        --------
+
+        """
         self.tempGap2 = tempGap2
         self.tempGap1 = tempGap1
         self.logQK = logQK
@@ -123,6 +182,23 @@ class Equilibrium(object):
 
 class pHDependenceType1(object):
     def __init__(self, ph1, slope1, ph2, slope2):
+        """Initialization of Parameters (Reaction rate dependence on pH)
+
+        Parameters
+        -----------
+        ph1 : float
+            pH of first mineral in reaction
+        slope1 : float
+            slope of first mineral in reaction
+        ph2 : float
+            pH of second mineral in reaction
+        slope2 : float
+            slope of second mineral in reaction
+
+        Returns
+        --------
+
[…]
