### 2024-04-07 acc1dff
remove model.rst
