### 2024-04-07 9b22be3
modify exclude patterns in conf.py
