+++ documentation/conf.py
@@ -27,7 +27,8 @@ extensions = ["sphinx.ext.todo", "sphinx.ext.viewcode", "sphinx.ext.autodoc", "s
               "sphinx.ext.napoleon", "sphinx.ext.intersphinx", "sphinx.ext.extlinks", 'sphinx_rtd_theme']
 
 templates_path = ['_templates']
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', 'documentation/constants.rst']
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', 'documentation/constants.rst', 'documentation/model.rst',
+                    'documentation/exceptions.rst']
 
 
 # -- Options for HTML output -------------------------------------------------
