### 2024-04-07 8a5c868
add docstrings to results
