+++ src/pytoughreact/results/multi_result_tough_react.py
@@ -32,6 +32,28 @@ from pytoughreact.results.result_tough_react import ResultReact
 class MultiResultReact(object):
     """ Class for retrieving results from multiple files for TOUGHREACT """
     def __init__(self, simulator_type, file_location, file_title, prop, **kwargs):
+        """Initialization of Parameters
+
+        Parameters
+        -----------
+        simulator_type :  list[string]
+            List of type of simulator being run. Can either be 'tmvoc', 'toughreact' or 'tough3'.
+            Should be toughreact for this class
+        file_location : list[string]
+            List of location of results file on system
+        file_title : list[string]
+            List of title or name of the file. Example is 'kddconc.tec'
+        prop : string
+            Prperty to be plotted. Example could be 'portlandite'
+        kwargs: dict
+            Extra property for processing. Takes in x_slice_value as a float to know at what
+            point to slice the data on the x axis.  
+
+
+        Returns
+        --------
+
+        """
         assert isinstance(file_location, list)
         assert isinstance(file_title, list)
         assert isinstance(prop, list)
@@ -45,14 +67,27 @@ class MultiResultReact(object):
         return 'Multiple Results from provided file locations and provided files for' + self.simulator_type
 
     def retrieve_data_multi_timeseries(self, grid_block_number, format_of_date='year'):
-        """ DataFrame to retrieve time and timeseries results from file """
+        """ Function that retrieves time and timeseries results from file
+
+        Parameters
+        -----------
+        grid_block_number :  int
+            The grid block number for which to retrieve the results
+        format_of_date : str
+            Provides information to the method on format of the date. For example. year, hour, min or seconds
+
+        Returns
+        --------
+        data_table : pd.Dataframe
+            Dataframe with requested output
+        """
         data_table = pd.DataFrame()
         for i in range(0, len(self.file_location)):
             tough_data = ResultReact(self.simulator_type, self.file_location[i], self.file_title[i])
             print(self.file_location[i])
             os.chdir(self.file_location[i])
             result_data = tough_data.get_timeseries_data(self.prop[0], grid_block_number)
-            time_data = tough_data.convert_times(format_of_date='year')
+            time_data = tough_data.convert_times(format_of_date=format_of_date)
             time_data_label = 'time' + str(i)
             result_data_label = 'result' + str(i)
             data_table[time_data_label] = pd.Series(time_data)
@@ -60,7 +95,20 @@ class MultiResultReact(object):
         return data_table
 
     def retrieve_data_multi_file_fixed_time(self, direction, time):
-        """ DataFrame to retrieve time and coordinate results from file """
+        """ DataFrame to retrieve time and coordinate results from file
+
+        Parameters
+        -----------
+        direction :  string
+            Direction of retrieval. Can be 'X', 'Y' or 'Z'
+        time : float
+            Time in which the data should be retrieved.
+
+        Returns
+        --------
+        data_table : pd.Dataframe
+            Dataframe with requested output
+        """
         data_table = pd.DataFrame()
         for i in range(0, len(self.file_location)):
             tough_data = ResultReact(self.simulator_type, self.file_location[i], self.file_title[i])
@@ -75,7 +123,22 @@ class MultiResultReact(object):
         return data_table
 
     def retrieve_data_multi_file_fixed_time_layer(self, direction, time, layer_num):
-        """ DataFrame to retrieve distance and results from file """
+        """ DataFrame to retrieve distance and results from file
+
+        Parameters
+        -----------
+        direction :  string
+            Direction of retrieval. Can be 'X', 'Y' or 'Z'
+        time : float
+            Time in which the data should be retrieved.
+        layer_num: int
+            Layer number in which to retrieve data
+
+        Returns
+        --------
[…]
