+++ src/pytoughreact/results/result_multiple.py
@@ -29,18 +29,32 @@ from pytoughreact.plotting.plot_multiple_files_routine import PlotMultiFiles
 class FileReadMultiple(object):
     """
     Class for processing multiple file results
-    file_locations (list of strings) - specifies the location of the files on the system
-    file_titles (list of strings) - gives the title of the file e.g 'kdd.conc' or 'OUTPUT.csv
-    simulator_type (string) can either be toughreact, tmvoc or tough3
-    props (list of strings) -  are the properties to be plotted
-
-    **kwargs
-    x_slice value (integer) - if the plot should be sliced a the  x axis
-    per_file -  (boolean) - if the plot should be made per file and not per property
-    title (list of strings) - title of each of the plots
     """
 
     def __init__(self, simulator_type, file_locations, file_titles, props, **kwargs):
+        """Initialization of Parameters
+
+        Parameters
+        -----------
+        simulator_type :  string
+            List of type of simulator being run. Can either be 'tmvoc', 'toughreact' or 'tough3'.
+            Should be tough3 for this class
+        file_location : list[string]
+            List of location of results file on system
+        file_title : list[string]
+            List of title or name of the file. Example is 'kddconc.tec' or 'OUTPUT.csv'
+        prop : list[string]
+            Properties to be plotted. Example could be 'portlandite'
+        kwargs: dict
+            x_slice value (integer) - if the plot should be sliced on the  x axis
+            per_file -  (boolean) - if the plot should be made per file and not per property
+            title (list of strings) - title of each of the plots
+
+
+        Returns
+        --------
+
+        """
         assert isinstance(file_locations, list)
         assert isinstance(file_titles, list)
         self.file_locations = file_locations
