### 2024-04-07 2160dba
update doc string for results
