### 2024-04-08 cc9950a
update docstrings from results folder
