+++ src/pytoughreact/results/result_tough_3.py
@@ -34,6 +34,25 @@ from pytoughreact.utilities.t2_utilities import t2Utilities
 class ResultTough3(object):
     """ Class for processing results from Tough3 """
     def __init__(self, simulatortype, filelocation, filetitle=None, **kwargs):
+        """Initialization of Parameters
+
+        Parameters
+        -----------
+        simulator_type :  string
+            Type of simulator being run. Can either be 'tmvoc', 'toughreact' or 'tough3'.
+            Should be tough3 for this class
+        file_location : string
+            Location of results file on system
+        file_title : string
+            Title or name of the file. Example is 'kddconc.tec' or 'OUTPUT.csv'
+        kwargs: dict
+            1) generation (string) - if generation data exists in the results.
+
+
+        Returns
+        --------
+
+        """
         if filelocation is None:
             self.filelocation = os.getcwd()
         else:
@@ -48,7 +67,18 @@ class ResultTough3(object):
         return 'Results from ' + self.filelocation + ' in ' + self.filetitle + ' for ' + self.simulatortype
 
     def read_file(self):
-        """ Read file """
+        """ Read file specified in file_location and file_title
+
+        Parameters
+        -----------
+
+
+        Returns
+        --------
+        file_as_list : list
+            Results from file as list
+
+        """
         os.chdir(self.filelocation)
         with open(self.filetitle) as csv_file:
             csv_reader = csv.reader(csv_file, delimiter=',', quotechar='"')
@@ -58,50 +88,90 @@ class ResultTough3(object):
         return self.file_as_list
 
     def get_times(self):
-        """
-        get times stored for duration of the simulation
-        :return: a list of all times
+        """ Get times stored for duration of the simulation
+
+        Parameters
+        -----------
+        grid_block_number :  int
+            The grid block number for which to retrieve the results
+        format_of_date : str
+            Provides information to the method on format of the date. For example. year, hour, min or seconds
+
+        Returns
+        --------
+        unprocessed_time_data : list
+            Time data directly from file without processing.
         """
         self.read_file()
         time = []
-        timeraw = []
+        unprocessed_time_data = []
         if self.generation is True:
             for i in range(1, len(self.file_as_list)):
-                timeraw.append(float(self.file_as_list[i][0]))
+                unprocessed_time_data.append(float(self.file_as_list[i][0]))
         else:
             for i in range(len(self.file_as_list)):
                 if len(self.file_as_list[i]) == 1:
                     time.append(self.file_as_list[i])
             for i in range(len(time)):
                 interim = time[i][0].split()
-                timeraw.append(float(interim[2]))
-        return timeraw
+                unprocessed_time_data.append(float(interim[2]))
+        return unprocessed_time_data
 
     def convert_times(self, format_of_date):
-        """
-        convert time to desirable time e.g day, month, year
-        :param format_of_date: string of type 'day', 'month', 'year'
-        :return: a list of the time
+        """ Convert time to desirable format e.g day, month, year
+
+        Parameters
+        -----------
+        format_of_date : str
+            Provides information to the method on format of the date. For example. year, hour, min or seconds
+
+        Returns
+        --------
+        processed_time_data  : list
+            List of converted time
+
         """
         intermediate = self.get_times()
         utility_function = t2Utilities()
-        timeyear = utility_function.convert_times(intermediate, format_of_date)
-        return timeyear
+        processed_time_data = utility_function.convert_times(intermediate, format_of_date)
+        return processed_time_data
 
     def get_time_index(self):
-        """ Get Index of Time """
+        """ Get Index of Time
+
+        Parameters
+        -----------
+
+
[…]
