+++ documentation/utilities.rst
@@ -3,10 +3,10 @@ Utilities Class
 ++++++++++++++++++++++
 TOUGHREACT Utilities
 ++++++++++++++++++++++
-.. autoclass:: pytoughreact.constants.t2_tough_react_utilities.t2UtilitiesToughReact
+.. autoclass:: pytoughreact.utilities.t2_tough_react_utilities.t2UtilitiesToughReact
     :members:
 ++++++++++++++++++++++
 TOUGH3 Utilities
 ++++++++++++++++++++++
-.. autoclass:: pytoughreact.constants.t2_utilities.t2Utilities
+.. autoclass:: pytoughreact.utilities.t2_utilities.t2Utilities
     :members:
