### 2024-04-08 c42dd96
update docstring for utilities
