+++ src/pytoughreact/utilities/t2_utilities.py
@@ -31,49 +31,105 @@ class t2Utilities(object):
         pass
 
     def convert_times(self, arraylist, format_of_date):
-        """ Convert times to second/minute/hour/day/year """
-        intermediate = arraylist
-        timeyear = []
+        """ Convert times to second/minute/hour/day/year
+
+        Parameters
+        -----------
+        arraylist :  list
+            Array of the time to be converted
+        format_of_date : str
+            Provides information to the method on format of the date. For example. year, hour, min or seconds
+
+        Returns
+        --------
+        processed_time_data : list
+            Time data after conversion
+        """
+        processing_data_array = arraylist
+        processed_time_data = []
         if format_of_date.lower() == 'year':
-            for i in range(len(intermediate)):
-                timeyear.append(intermediate[i] / 3.154e+7)
+            for i in range(len(processing_data_array)):
+                processed_time_data.append(processing_data_array[i] / 3.154e+7)
         elif format_of_date.lower() == 'day':
-            for i in range(len(intermediate)):
-                timeyear.append(intermediate[i] / 86400)
+            for i in range(len(processing_data_array)):
+                processed_time_data.append(processing_data_array[i] / 86400)
         elif format_of_date.lower() == 'hour':
-            for i in range(len(intermediate)):
-                timeyear.append(intermediate[i] / 3600)
+            for i in range(len(processing_data_array)):
+                processed_time_data.append(processing_data_array[i] / 3600)
         elif format_of_date.lower() == 'minute':
-            for i in range(len(intermediate)):
-                timeyear.append(intermediate[i] / 60)
+            for i in range(len(processing_data_array)):
+                processed_time_data.append(processing_data_array[i] / 60)
         elif format_of_date.lower() == 'second':
-            for i in range(len(intermediate)):
-                timeyear.append(intermediate[i])
+            for i in range(len(processing_data_array)):
+                processed_time_data.append(processing_data_array[i])
         else:
             raise ValueError("format can either be year, day, hour, minute or second")
-        return timeyear
+        return processed_time_data
 
-    def choplist(self, liste, number=3):
-        """ Reduce the lsit size """
-        global finallist
-        if isinstance(liste, list):
-            if len(liste) > 100:
-                liste = liste[0:len(liste):number]
-                self.choplist(liste)
+    def chop_list(self, input_list, step_increase=3):
+        """ Reduce the length of the list
+
+        Parameters
+        -----------
+        input_list :  list[float]
+            List for which its size is to be reduced
+        step_increase : int
+            Step increase to used in the length reduction
+
+        Returns
+        --------
+        final_processed_list : list
+            List data after reduction
+        """
+        global final_processed_list
+        if isinstance(input_list, list):
+            if len(input_list) > 100:
+                input_list = input_list[0:len(input_list):step_increase]
+                self.chop_list(input_list)
             else:
-                finallist = liste[0:len(liste):number]
-        return finallist
+                final_processed_list = input_list[0:len(input_list):step_increase]
+        return final_processed_list
 
-    def cutdata(self, time_data, resultdata, slicevalue):
-        """ Cut the result output to avoid too many points on a graph """
+    def trim_data_points(self, time_data, result_data, slice_value):
+        """ Trim the result output to avoid too many points on a graph
+
+        Parameters
+        -----------
+        time_data :  list[float]
+            Time data list
+        result_data :  list[float]
+            Result data list
+        slice_value : float
+            Number at which to trim the data
+
+        Returns
+        --------
+        time_data, result_data : list , list
+            List data after trimming
+
[…]
