+++ src/pytoughreact/utilities/t2_tough_react_utilities.py
@@ -34,23 +34,28 @@ class t2UtilitiesToughReact(object):
     This class prepares the output files from TOUGHREACT for plot visualizations and analysis
     """
 
-    def __init__(self, location, word, file2='MESH'):
-
-        """
+    def __init__(self, location, word, mesh_file_name='MESH'):
+        """ Initialization of Parameters
         An instance of this class takes in five parameters the last of which is optional;
 
-        location --> the current direction where the simulations have been carried out
-        destination ---> the directory containing PYTOUGH and its class which would be needed for
-        manipulations
-        filenames -> the result files to be transferred to the destination folder
-        word --> the word where the truncation in the MESH file is to begin. Typically this should be 'CONNE'
+        Parameters
+        -----------
+        location : string
+            The current file location where the simulations have been carried out
+        word : string
+            The word where the truncation in the MESH file is to begin. Typically this should be 'CONNE'
+        mesh_file_name : string
+            The name of the mesh file
+
+        Returns
+        --------
         """
 
         self.location = location
         self.word = word
-        self.file2 = file2
+        self.file2 = mesh_file_name
 
-    def copyfile(self, filename, destination):
+    def copy_file(self, filename, destination):
         """ This method copies single file from the location to the destination folder. it takes in a a single argument
 
         Parameters
@@ -62,6 +67,7 @@ class t2UtilitiesToughReact(object):
 
         Returns
         --------
+
         """
         # copy specific file
         src_files = os.listdir(self.location)
@@ -71,7 +77,7 @@ class t2UtilitiesToughReact(object):
                 if (os.path.isfile(full_file_name)):
                     shutil.copy(full_file_name, destination)
 
-    def copyallfiles(self, filenames):
+    def copy_all_files(self, filenames):
         """This method copies all files given in the instance of the class to the destination folder. It makes use
         of the copyfile() method in achieving this
 
@@ -87,10 +93,10 @@ class t2UtilitiesToughReact(object):
         # copy all files
         for i in range(0, len(filenames)):
             a = filenames[i]
-            self.copyfile(a)
+            self.copy_file(a)
         print('...copying files...')
 
-    def findword(self):
+    def find_word(self):
         """ This method finds the word where the truncation of the MESH file is to occur.
 
         Parameters
@@ -110,7 +116,7 @@ class t2UtilitiesToughReact(object):
                     return point1
         myFile.close()
 
-    def sliceofffile(self):
+    def slice_off_file(self):
         """ This method slices off all parameters below the word stated in the instance of the class
 
         Parameters
@@ -132,14 +138,14 @@ class t2UtilitiesToughReact(object):
         f.truncate(0)
         f.close()
         os.remove("test.txt")
-        point1 = self.findword()
+        point1 = self.find_word()
         with open("test2.txt", "w") as f1:
             with open(self.file2, "r") as text_file:
                 for line in itertools.islice(text_file, 1, point1 - 2):
                     f1.write(line)
         f1.close()
 
-    def sliceoffline(self):
+    def slice_off_line(self):
         """ This method slices off all grid parameter such as the volume, distance betweeen grids as stated in
         the TOUGHREACT flow.inp file
 
@@ -155,7 +161,7 @@ class t2UtilitiesToughReact(object):
             list of grids
 
         """
-        self.sliceofffile()
+        self.slice_off_file()
         with open('test2.txt') as thefile:
             lines = thefile.readlines()
             output = []
@@ -165,13 +171,12 @@ class t2UtilitiesToughReact(object):
                 output.append(b)
 
         return output
-        thefile.close()
 
-    def writetofile(self):
+    def write_to_file(self):
[…]

[…]
