### 2024-04-08 0172ef3
update pyproject.toml and dependencies.rst
