+++ documentation/dependencies.rst
@@ -5,5 +5,9 @@ Dependencies
 1. VTK
 2. Scipy
 3. Numpy
+4. Matplotlib
+5. pandas
+6. PyTOUGH
+7. flake8
 
 
