### 2024-04-08 temmy222/PyTOUGHREACT#7 @temmy222
Hi @frank1010111 . Thanks for your review. I believe I have fixed most of the issues you raised on the documentation. Please let me know if I missed any. I will begin fixing the issues you raised with the tutorials immediately.
