### 2024-04-10 temmy222/PyTOUGHREACT#9 @blakeaw
JOSS Review - Link to License in README
This Issue has been raised as a part of the PyTOUGHREACT JOSS Review at: https://github.com/openjournals/joss-reviews/issues/6486

Hi @temmy222,

The license link in the README points to the MIT license template on Choose a License: https://choosealicense.com/licenses/mit/
Is this on purpose, or should that link to your copy of the license stored in your repository?
