### 2024-04-10 temmy222/PyTOUGHREACT#10 @blakeaw
JOSS Review - Python requirement version in docs
This Issue has been raised as a part of the PyTOUGHREACT JOSS Review at: https://github.com/openjournals/joss-reviews/issues/6486

The Python version is given as >=3.7 in the README, but this is missing from the Dependencies section of the docs: https://pytoughreact.readthedocs.io/en/latest/dependencies.html

Instead, I see the following on my end: ```Python: pandas <https://pandas.pydata.org/> _ library```
Should that be something like ```Python: 3.7 or greater```? Can you double-check?
