### 2024-04-10 temmy222/PyTOUGHREACT#11 @blakeaw
JOSS Review - Non-code contributing author
This Issue has been raised as a part of the PyTOUGHREACT JOSS Review at: https://github.com/openjournals/joss-reviews/issues/6486

Can you clarify the contributions of the corresponding author? I see @frank1010111 asked about this already in the review thread: https://github.com/openjournals/joss-reviews/issues/6486#issuecomment-1998824404. Could you let me know if you've already addressed this? If so, can you point me to the relevant response? Thanks.
