### 2024-04-10 temmy222/PyTOUGHREACT#10 @blakeaw
I also see that there are specific versions for the other dependencies in the pyproject.toml: https://github.com/temmy222/PyTOUGHREACT/blob/master/pyproject.toml ; e.g., `numpy>=1`

However, these requirements aren't specified in the README or docs. It could be worth including those details in the Dependencies section of the docs.
