### 2024-04-10 temmy222/PyTOUGHREACT#12 @blakeaw
JOSS Review - Tutorial: library imports
This Issue has been raised as a part of the PyTOUGHREACT JOSS Review at: https://github.com/openjournals/joss-reviews/issues/6486

In the example shown in the Tutorial (https://pytoughreact.readthedocs.io/en/latest/tutorial.html) there are many library imports.

```python
import os
from mulgrids import mulgrid
from pytoughreact.writers.react_writing import t2react
from pytoughreact.wrapper.reactgrid import t2reactgrid
from pytoughreact.wrapper.reactzone import t2zone
from pytoughreact.chemical.chemical_composition import PrimarySpecies, WaterComp, Water, ReactGas
from pytoughreact.chemical.mineral_composition import MineralComp
from pytoughreact.chemical.mineral_zone import MineralZone
from pytoughreact.constants.default_minerals import get_kinetics_minerals, get_specific_mineral
from pytoughreact.writers.chemical_writing import t2chemical
from pytoughreact.chemical.perm_poro_zone import PermPoro, PermPoroZone
from pytoughreact.writers.solute_writing import t2solute
from t2grids import rocktype
```

1. You might add some additional explanation and context describing what each import is for. Maybe just as comments in the code block? I think that could be helpful for a new user.
2. There are also a couple of libraries that aren't from `pytoughreact` and it isn't immediately clear where they come from, including `mulgrids` and `t2grids`. Are these part of PyTOUGH? You might also note this for clarity.
3. As there are many imports needed, maybe a shorthand like `from pytoughreact import *` could be helpful, at least for more advanced users. I know many of these are included in `pytoughreact/__init__.py`, but there are also several such as the `writers` and `wrapper` objects that aren't imported in the `__init__.py`.
