### 2024-04-10 temmy222/PyTOUGHREACT#7 @blakeaw
Hi @temmy222, I'm not sure if you have addressed this already, but I wanted to note that I agree with @frank1010111's comment about insufficient information in most docstrings:

> Most of the docstrings provide insufficient information to use the method, class, or function.

I ran into this problem while trying to follow along with the tutorial. I thought maybe the API docs would help, but they are lacking descriptions of the different objects (or have only very short ones). In particular, for each class, I'd recommend describing the object it represents along with some context on its purpose in the simulations.

If you have addressed this already please let me know, as any changes don't seem to be reflected in the latest version of the docs on readthedocs.
