### 2024-04-10 temmy222/PyTOUGHREACT#13 @blakeaw
JOSS Review - PyTOUGHREACT version in version.py isn't updated
This Issue has been raised as a part of the PyTOUGHREACT JOSS Review at: https://github.com/openjournals/joss-reviews/issues/6486

It looks like the version number for pytoughreact in `src/pytoughreact/version.py` hasn't been updated to reflect the newest version number of 1.0.4. It is still listed there as 0.0.15, so `print(pytoughreact.__version__` still outputs `0.0.15`.
