+++ src/pytoughreact/chemical/biomass_composition.py
@@ -44,23 +44,23 @@ class Gas(object):
         self.index = index
         self.name = name
 
-    def addToProcess(self, process, Uptake, Ks=None, Kc=None, Knc=None, Kh=None):
+    def add_to_process(self, process, uptake, ks=None, kc=None, knc=None, kh=None):
         """ Add gas component to the process
 
         Parameters
         -----------
         process :  Process
             This should be a Process class with all properties of the process
-        Uptake : int
+        uptake : int
             uptake coefficient of gas component in particular process with respect to 1 mole of
             degraded primary substrate (mole component / mole substrate).
-        Ks: float
+        ks: float
             Substrate degradation rate
-        Kc: float
+        kc: float
             Competitive inhibition rate
-        Knc: float
+        knc: float
             Non Competitive inhibition rate
-        Kh: float
+        kh: float
             Haldane inhibition rate
 
         Returns
@@ -70,42 +70,42 @@ class Gas(object):
         process : Process
             Updated Process with new parameters
         """
-        output = {self: [Uptake, Ks, Kc, Knc, Kh]}
-        if Kc is not None:
+        output = {self: [uptake, ks, kc, knc, kh]}
+        if kc is not None:
             process.NumOfCompetiting += 1
-        if Knc is not None:
+        if knc is not None:
             process.NumOfNonCompetiting += 1
-        if Kh is not None:
+        if kh is not None:
             process.NumOfHaldane += 1
         process.componentParams = output
         process.allProcesses.append(output)
         return output, process
 
 
-class Water_Bio(object):
+class WaterBio(object):
     """Rock type"""
 
     def __init__(self, name='H2O', index=1):
         self.index = index
         self.name = name
 
-    def addToProcess(self, process, Uptake, Ks=None, Kc=None, Knc=None, Kh=None):
+    def add_to_process(self, process, uptake, ks=None, kc=None, knc=None, kh=None):
         """ Add water component to the process
 
         Parameters
         -----------
         process :  Process
             This should be a Process class with all properties of the process
-        Uptake : int
+        uptake : int
             uptake coefficient of water component in particular process with respect to 1 mole of
             degraded primary substrate (mole component / mole substrate).
-        Ks: float
+        ks: float
             Substrate degradation rate
-        Kc: float
+        kc: float
             Competitive inhibition rate
-        Knc: float
+        knc: float
             Non Competitive inhibition rate
-        Kh: float
+        kh: float
             Haldane inhibition rate
 
         Returns
@@ -115,12 +115,12 @@ class Water_Bio(object):
         process : Process
             Updated Process with new parameters
         """
-        output = {self: [Uptake, Ks, Kc, Knc, Kh]}
-        if Kc is not None:
+        output = {self: [uptake, ks, kc, knc, kh]}
+        if kc is not None:
             process.NumOfCompetiting += 1
-        if Knc is not None:
+        if knc is not None:
             process.NumOfNonCompetiting += 1
-        if Kh is not None:
+        if kh is not None:
             process.NumOfHaldane += 1
         process.componentParams = output
         process.allProcesses.append(output)
@@ -130,52 +130,57 @@ class Water_Bio(object):
 class BaseComponent(object):
     """Rock type"""
 
-    def __init__(self, name=None, critTemp=None, critPres=None, critComp=None, acentricFactor=None, dipoleMoment=None,
-                 boilPoint=None, vapPressA=None, vapPressB=None, vapPressC=None, vapPressD=None,
[…]
