### 2024-04-11 1e7e372
update dependencies in docs
