### 2024-04-11 c7c8e94
update dependencies.rst
