+++ documentation/dependencies.rst
@@ -2,11 +2,11 @@ Dependencies
 ================
 **Python**: Python > 3.7
 
-1. VTK $\geq$ 1
-2. Scipy >=1
-3. Numpy >=1
-4. Matplotlib >=3
-5. pandas >=1
+1. VTK >= 1
+2. Scipy >= 1
+3. Numpy >= 1
+4. Matplotlib >= 3
+5. pandas >= 1
 6. PyTOUGH
 7. flake8
 
