+++ src/pytoughreact/chemical/biomass_composition.py
@@ -25,9 +25,37 @@ SOFTWARE.
 
 
 class Biomass(object):
-    """Rock type"""
+    """Biomass type"""
 
     def __init__(self, index, name, init_conc, min_conc, max_temp, death_rate, inhibition_constant):
+        """Initialization of Parameters
+
+        Parameters
+        -----------
+        index :  int
+            Serial index of biomass
+        name : string
+            Name of biomass
+        init_conc: float
+            Initial concentration (valid for the entire simulation grid) in the aqueous phase of
+            microbial population (kg biomass / kg aqueous phase)
+        min_conc: float
+            Minimum concentration in the aqueous phase of microbial population
+            enforced during the simulation (kg biomass / kg aqueous phase)
+        max_temp: float
+            Maximum temperature for the calculation of temperature inhibition function in the
+            substrate degradation rate equation (°C)
+        death_rate: float
+            Death rate constant, or maintenance constant, for the microbial population
+            (s-1)
+        inhibition_constant: float
+            Inhibition constant for biomass growth of microbial population # IB (kg biomass
+            kg aqueous phase).
+
+        Returns
+        --------
+
+        """
         self.index = index
         self.death_rate = death_rate
         self.max_temp = max_temp
@@ -41,6 +69,19 @@ class Gas(object):
     """Rock type"""
 
     def __init__(self, name, index):
+        """Initialization of Parameters
+
+        Parameters
+        -----------
+        index :  int
+            Serial index of gas
+        name : string
+            Name of gas
+
+        Returns
+        --------
+
+        """
         self.index = index
         self.name = name
 
@@ -86,6 +127,19 @@ class WaterBio(object):
     """Rock type"""
 
     def __init__(self, name='H2O', index=1):
+        """Initialization of Parameters
+
+        Parameters
+        -----------
+        index :  int
+            Serial index of water
+        name : string
+            Name of water
+
+        Returns
+        --------
+
+        """
         self.index = index
         self.name = name
 
@@ -130,19 +184,94 @@ class WaterBio(object):
 class BaseComponent(object):
     """Rock type"""
 
-    def __init__(self, name=None, critical_temperature=None, critical_pressure=None, critical_composition=None,
+    def __init__(self, name=None, critical_temperature=None, critical_pressure=None, critical_compressibility=None,
                  acentric_factor=None, dipole_moment=None,
                  boiling_point=None, vapor_pressure_a=None, vapor_pressure_b=None,
                  vapor_pressure_c=None, vapor_pressure_d=None,
                  molecular_weight=None, heat_capacity_constant_a=None, heat_capacity_constant_b=None,
                  heat_capacity_constant_c=None, heat_capacity_constant_d=None,
-                 liquid_density=None, reference_temp_for_density=None, reference_binary_difference=None,
-                 reference_temperature_for_difference=None, exponent_chemical_difference=None,
+                 liquid_density=None, reference_temp_for_density=None, reference_binary_diffusivity=None,
+                 reference_temperature_for_diffusivity=None, exponent_chemical_diffusivity=None,
                  liquid_viscosity_constant_a=None, liquid_viscosity_constant_b=None, liquid_viscosity_constant_c=None,
                  liquid_viscosity_constant_d=None, liquid_critical_volume=None,
                  liquid_chemical_solubility_a=None, liquid_chemical_solubility_b=None,
                  liquid_chemical_solubility_c=None, liquid_chemical_solubility_d=None,
-                 carbon_part_coefficient=None, fractional_carbon=None, decay_constant=None):
+                 carbon_partition_coefficient=None, fractional_organic_carbon=None, decay_constant=None):
+        """Initialization of Parameters
+
[…]
