### 2024-04-12 #6486 @frank1010111
So far my review is going well, but one hang-up is testing the functionality of the software. At the moment, I have no way to verify that PyTOUGHREACT produces valid TOUGH scripts when given valid inputs. The automated tests do not check outputs to ensure they conform to any particular grammar. That makes it really hard to check that particular box.
