+++ examples/example_bio/run_bio.py
@@ -44,11 +44,11 @@ simulation_time = 100 * year
 
 length = 1000.
 x_block = 10
-yblock = 1
-zblock = 5
+y_block = 1
+z_block = 5
 dx = [length / x_block] * x_block
 dy = [1.0]
-dz = [5] * zblock
+dz = [5] * z_block
 geo = mulgrid().rectangular(dx, dy, dz, origin=[0, 0, -95])
 geo.write('geom.dat')
 
@@ -117,7 +117,7 @@ well = 'wl '
 compo = ['COM3']
 direction = 'x'
 duration = [0, 1 * year, 101 * year]
-# duration = np.linspace(0, simtime * 2, 7)
+# duration = np.linspace(0, simulation_time * 2, 7)
 rate = np.array([1.00e-2, 0, 0])
 rate_o2 = [1.00e-03, 0, 0]
 energy = [5, 5, 5]
@@ -125,7 +125,7 @@ energy = [5, 5, 5]
 if direction == 'x':
     j = 0
     for i in range(0, x_block):
-        # for i in range(xblock * (zblock - 1), xblock * (zblock)):
+        # for i in range(x_block * (z_block - 1), x_block * (z_block)):
         for component in compo:
             if component == 'COM2':
                 gen = t2generator(name=well + str(i), block=bio.grid.blocklist[i].name, type=component,
@@ -143,6 +143,6 @@ if direction == 'x':
 
 # ____________________________________RUN SIMULATION_______________________________________________________________
 
-runlocation = os.getcwd()
-bio.write('INFILE', runlocation=os.getcwd())
-bio.run(simulator='tmvoc', runlocation='')
+run_location = os.getcwd()
+bio.write('INFILE', run_location=os.getcwd())
+bio.run(simulator='tmvoc', run_location='')
