### 2024-04-18 6696ffc
delete checkmarx.yml
