+++ src/pytoughreact/__init__.py
@@ -8,7 +8,7 @@ from pytoughreact.chemical.mineral_composition import MineralComp
 from pytoughreact.chemical.mineral_description import Mineral
 from pytoughreact.chemical.perm_poro_zone import PermPoro, PermPoroZone
 from pytoughreact.chemical.mineral_zone import MineralZone
-from pytoughreact.writers.bio_writing import T2Bio
+from pytoughreact.writers.bio_writing import T2Bio, T2BioParser
 from pytoughreact.chemical.biomass_composition import Component, Gas, WaterBio, Biomass, BaseComponent, Solids
 from pytoughreact.chemical.bio_process_description import Process, BIODG
 from pytoughreact.plotting.plot_single import PlotSingle
@@ -16,7 +16,6 @@ from pytoughreact.wrapper.react_data import ReactData
 from pytoughreact.wrapper.reactblock import T2Block
 from pytoughreact.wrapper.reactgrid import T2ReactGrid
 from pytoughreact.wrapper.reactzone import T2Zone
-from pytoughreact.writers.bio_writing import T2Bio, T2BioParser
 from pytoughreact.writers.chemical_writing import T2Chemical, T2ChemicalData
 from pytoughreact.writers.react_writing import T2React, T2ReactParser, T2ExtraPrecisionDataParser
 from pytoughreact.writers.solute_writing import T2Solute, T2SoluteParser
