### 2024-04-28 temmy222/PyTOUGHREACT#12 @temmy222
Hi

I have modified the code to correct these concerns

1) Explanation via comments added
2) Yes, these are part of PyTOUGH and have been added in the comments
3) Yes, this shorthand can be used. Or the individual imports can be used. The individual imports were used to help beginner users understand where what is being imported is coming from.
