### 2024-04-28 temmy222/PyTOUGHREACT#7 @temmy222
Hi @frank1010111. Thanks for your review

Here are comments to your first comments on this issue

1) Link to PyTOUGH has been corrected
2) Documentation updated to correct all mentioned points
3) Contributions updated
4) Dependencies same as pyproject.toml
5) License updated to show MIT everywhere
6) Paper content itself is less than 1000 words. The examples are added which make it appear to be more than 1000 words
