+++ docs/publication/paper.md
@@ -62,19 +62,20 @@ Two broad examples are shown in the next two sections to show how the package ca
 
 import os
 from mulgrids import mulgrid
-from pytoughreact.writers.react_writing import t2react
-from pytoughreact.wrapper.reactgrid import t2reactgrid
-from pytoughreact.wrapper.reactzone import t2zone
+from pytoughreact.writers.react_writing import T2React
 from pytoughreact.chemical.chemical_composition import PrimarySpecies, WaterComp, Water, ReactGas
+from pytoughreact.wrapper.reactgrid import T2ReactGrid
+from pytoughreact.wrapper.reactzone import T2Zone
 from pytoughreact.chemical.mineral_composition import MineralComp
 from pytoughreact.chemical.mineral_zone import MineralZone
 from pytoughreact.constants.default_minerals import get_kinetics_minerals, get_specific_mineral
-from pytoughreact.writers.chemical_writing import t2chemical
+from pytoughreact.chemical.mineral_description import Mineral
+from pytoughreact.writers.chemical_writing import T2Chemical
 from pytoughreact.chemical.perm_poro_zone import PermPoro, PermPoroZone
-from pytoughreact.writers.solute_writing import t2solute
+from pytoughreact.writers.solute_writing import T2Solute
 from t2grids import rocktype
 
-# __________________________________FLOW.INP____________________________________________
+#__________________________________FLOW.INP____________________________________________
 length = 0.1
 nblks = 1
 dx = [length / nblks] * nblks
@@ -83,13 +84,13 @@ dz = [0.5] * 1
 geo = mulgrid().rectangular(dx, dy, dz)
 geo.write('geom.dat')
 
-react = t2react()
+react = T2React()
 react.title = 'Reaction example'
 
 react.multi = {'num_components': 1, 'num_equations': 1, 'num_phases': 2,
                'num_secondary_parameters': 6}
 
-react.grid = t2reactgrid().fromgeo(geo)
+react.grid = T2ReactGrid().fromgeo(geo)
 
 react.parameter.update(
     {'print_level': 4,
@@ -110,7 +111,8 @@ react.grid.add_rocktype(sand)
 for blk in react.grid.blocklist[0:]:
     blk.rocktype = react.grid.rocktype[sand.name]
 
-zone1 = t2zone('zone1')
+
+zone1 = T2Zone('zone1')
 
 react.grid.add_zone(zone1)
 
@@ -121,7 +123,7 @@ react.start = True
 
 react.write('flow.inp')
 
-# ____________________________________CHEMICAL.INP______________________________________
+#____________________________________CHEMICAL.INP________________________________________
 h2o = PrimarySpecies('h2o', 0)
 h = PrimarySpecies('h+', 0)
 na = PrimarySpecies('na+', 0)
@@ -135,7 +137,7 @@ al = PrimarySpecies('al+3', 0)
 fe = PrimarySpecies('fe+2', 0)
 hs = PrimarySpecies('hs-', 0)
 
-all_species = [h2o, h, na, cl, hco3, ca, so4, mg, h4sio4, al, fe, hs]
+all_species = [h2o, h,na, cl, hco3, ca, so4, mg, h4sio4, al, fe, hs]
 
 h2o_comp1 = WaterComp(h2o, 1, 1.0000E+00, 1.000000E+00)
 h_comp1 = WaterComp(h, 1, 1E-7, 1E-7)
@@ -152,7 +154,7 @@ hs_comp1 = WaterComp(hs, 1, 1E-10, 1E-10)
 
[…]
