### 2024-04-28 temmy222/PyTOUGHREACT#11 @temmy222
Hi @blakeaw. Yes, the corresponding author was responsible for conceptualization, paper revision, supervision and funding. The corresponding author didn't write any of the code.
