### 2024-04-28 24c9781
:memo: fix README usage example
