### 2024-04-28 8c90c4e
:memo: fix solute and react.run in README usage example
