### 2024-04-28 temmy222/PyTOUGHREACT#7 @frank1010111
Hi Temmy,

Thank you for these changes. I am seeing improvement in the documentations.

Unfortunately, there have also been some breakages in the examples. I am trying to get things working again, as documented here: https://github.com/temmy222/PyTOUGHREACT/pull/17

Is this the [thddem.dat](https://github.com/temmy222/PyTOUGHREACT/blob/f509b58b04668f46f4f07a650d5c0eafb0f99ee2/thddem.dat) file that the code seems to need? There doesn't appear to be one in the current repository, so I had to go searching through the history.
