+++ documentation/developer.rst
@@ -28,18 +28,8 @@ A new page to the documentation is added as a .rst file to the documentation fol
 added to the index.rst file. Examples can be found in the index.rst file for how to add a new file
 to it.
 
-Updates to developer documentation are automatically upladed to https://pytoughreact.readthedocs.io/en/master/developer.html
+Updates to developer documentation are automatically uploaded to https://pytoughreact.readthedocs.io/en/master/developer.html
 
-Important links
-~~~~~~~~~~~~~~~~~~~~
-1. https://www.sphinx-doc.org/en/master/tutorial/automatic-doc-generation.html
-2. https://www.sphinx-doc.org/en/master/tutorial/getting-started.html#creating-the-documentation-layout
-3. https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html#literal-blocks
-4. https://sphinx-rtd-theme.readthedocs.io/en/stable/installing.html 
-5. https://sphinx-tutorial.readthedocs.io/step-3/
-6. https://docutils.sourceforge.io/docs/user/rst/quickref.html#block-quotes 
-7. https://www.youtube.com/watch?v=BWIrhgCAae0&ab_channel=LearnProgrammingwithJoel
-8. https://www.sphinx-doc.org/en/master/usage/restructuredtext/roles.html#cross-referencing-arbitrary-locations
 
 
 Test
