### 2024-04-28 temmy222/PyTOUGHREACT#7 @temmy222
Thank you for the fix @frank1010111. Yes, that is the required thddem.dat. I removed it from the repo because it seems to have been changed. Users can specify their database themselves
