### 2024-04-28 temmy222/PyTOUGHREACT#18 @frank1010111
JOSS review: JOSS paper
## Is your feature request related to a problem? Please describe.
This is part of an ongoing [JOSS review](https://github.com/openjournals/joss-reviews/issues/6486), specifically about the paper.

## Describe the solution you'd like

As outlined in the [submission guidelines provided to authors](https://github.com/openjournals/joss/blob/main/docs/submitting.md#what-should-my-paper-contain), the JOSS paper (the compiled PDF associated with this submission) should only include:

- [x] A list of the authors of the software and their affiliations.
- [ ] A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
- [x] A clear statement of need that illustrates the purpose of the software.
- [ ] A description of how this software compares to other commonly-used packages in this research area.
- [ ] Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it.
- [x] A list of key references including a link to the software archive.

```{important}
Note the paper *should not* include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.
```

### Aligning the manuscript with the guidelines

1. I consider the architecture and examples to be software documentation and think they should be moved from the paper to the sphinx-generated documentation. They're great, just not meant for the paper
2. At the moment, the introduction is a blend of a statement of need, purpose, and description of other tools, all in one paragraph. Ideally, it would be broken up into at least three paragraphs:
    - "PyTOUGHREACT is..." - high level functionality and purpose
    -  "Other tools working with TOUGH include..." - comparison to other packages, such as PyTOUGH, PetraSim, TOUGH2VIEWER, etc
    - "PyTOUGHREACT has enabled us to..." - list the research projects you've completed or started thanks to your tool.
