### 2024-04-30 temmy222/PyTOUGHREACT#8 @matthewfeickert
> The idea of the tests is not to run the simulations or check the outputs but to check that the software writes to the required file and can read all values from the files into the classes in the software.

@temmy222 Do the tests actually test the software at all? If they only test reading and writing this is not enough.
