### 2024-05-01 #6486 @matthewfeickert
> The automated tests do not check outputs to ensure they conform to any particular grammar. That makes it really hard to check that particular box.

@frank1010111 Can you check if https://github.com/temmy222/PyTOUGHREACT/issues/8 has been properly resolved to address this concern? It appears the issue was closed without addressing adding functionality testing. If that's correct it should be reopened.
