+++ test/test_t2reactgrid.py
@@ -1,0 +1,481 @@
+import os
+import sys
+from mulgrids import mulgrid
+from pytoughreact.chemical.kinetic_properties import PHDependenceType2, Dissolution, Precipitation
+from pytoughreact.chemical.mineral_description import Mineral
+from pytoughreact.chemical.chemical_composition import PrimarySpecies, WaterComp, Water, ReactGas
+from pytoughreact.chemical.mineral_composition import MineralComp
+from pytoughreact.chemical.mineral_zone import MineralZone
+from pytoughreact.chemical.perm_poro_zone import PermPoro, PermPoroZone
+from pytoughreact.writers.react_writing import T2React
+from pytoughreact.writers.solute_writing import T2Solute
+from pytoughreact.writers.chemical_writing import T2Chemical
+from pytoughreact.wrapper.reactzone import T2Zone
+from pytoughreact.wrapper.reactgrid import T2ReactGrid
+from pytoughreact.wrapper.reactblock import T2Block
+from pytoughreact.results.t2result import T2Result
+from t2data import rocktype
+
+
+class T2ReactGridTestCase():
+    def get_specific_mineral(self, mineral_name):
+        calcite_ph = PHDependenceType2(5.0119e-01, 14.4, 1, 'h+', 1.0)
+        dissolution_calcite = Dissolution(1.5488e-06, 2, 1, 1, 23.5, 0, 0, 0)
+        dissolution_calcite.ph_dependence = [calcite_ph]
+        precipitation_calcite = Precipitation(1.5488e-06, 0, 1, 1, 23.5, 0, 0, 0, 1.0E-5, 0, 0, 0, 0)
+        calcite = Mineral('Calcite', 1, 3, 0, 0)
+        calcite.dissolution = [dissolution_calcite]
+        calcite.precipitation = [precipitation_calcite]
+
+        dissolution_portlandite = Dissolution(2.18e-08, 0, 1, 1, 74.9)
+        precipitation_portlandite = Precipitation(2.18e-08, 0, 1, 1, 0, 0, 0, 0, 1e-6, 0, 0, 0, 0)
+        portlandite = Mineral('Portlandite', 1, 3, 0, 0)
+        portlandite.dissolution = [dissolution_portlandite]
+        portlandite.precipitation = [precipitation_portlandite]
+
+        dissolution_c3fh6 = Dissolution(1.60e-018, 0, 1, 1, 0.0)
+        precipitation_c3fh6 = Precipitation(1.60e-018, 0, 1, 1, 0, 0, 0, 0, 1e-6, 0, 0, 0, 0)
+        c3fh6 = Mineral('C3FH6', 1, 3, 0, 0)
+        c3fh6.dissolution = [dissolution_c3fh6]
+        c3fh6.precipitation = [precipitation_c3fh6]
+
+        dissolution_tobermorite = Dissolution(1.0e-012, 0, 1, 1, 0.0)
+        precipitation_tobermorite = Precipitation(1.0e-012, 0, 1, 1, 0, 0, 0, 0, 1e-6, 0, 0, 0, 0)
+        tobermorite = Mineral('Tobermorite(11A)', 1, 3, 0, 0)
+        tobermorite.dissolution = [dissolution_tobermorite]
+        tobermorite.precipitation = [precipitation_tobermorite]
+
+        dissolution_ettringite = Dissolution(1.2589e-012, 0, 1, 1, 0.0)
+        precipitation_ettringite = Precipitation(1.2589e-012, 0, 1, 1, 0, 0, 0, 0, 1e-6, 0, 0, 0, 0)
+        ettringite = Mineral('Ettringite', 1, 3, 0, 0)
+        ettringite.dissolution = [dissolution_ettringite]
[…]
