### 2024-05-01 temmy222/PyTOUGHREACT#19 @blakeaw
JOSS Review: pytest-cov needed to run tests with default pyproject.toml options
This Issue has been raised as a part of the PyTOUGHREACT JOSS Review at: https://github.com/openjournals/joss-reviews/issues/6486

When I first tried to run pytest, I got the following error:

> > ERROR: usage: __main__.py [options] [file_or_dir] [file_or_dir] [...]
> __main__.py: error: unrecognized arguments: --cov=pytoughreact
>   inifile: PyTOUGHREACT\pyproject.toml
>   rootdir: PyTOUGHREACT

After digging a little, I realized I needed pytest-cov for the `--cov=pytoughreact` option to be valid.

Since it is needed to generate the coverage report and run the tests without modifying the pyproject.toml first, I'd suggest adding pytest-cov installation to the notes in the following docs section: https://pytoughreact.readthedocs.io/en/latest/developer.html#test
