### 2024-05-01 temmy222/PyTOUGHREACT#20 @blakeaw
JOSS Review: Missing Community Guidelines
This Issue has been raised as a part of the PyTOUGHREACT JOSS Review at: https://github.com/openjournals/joss-reviews/issues/6486

The requirements for community guidelines are:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

It looks like number 1 is covered by the Contributing sections of the README and docs: https://pytoughreact.readthedocs.io/en/latest/contributing.html

However, I don't see any guidelines regarding the second two items `2) Report issues or problems with the software 3) Seek support`.
