### 2024-05-01 temmy222/PyTOUGHREACT#20 @blakeaw
You might also consider adding a CONTRIBUTING.md file to the repo that includes all the relevant community guidelines.
https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors#adding-a-contributing-file
