### 2024-05-02 #6486 @frank1010111
Hi @matthewfeickert :wave:
A few issues are still outstanding

## General checks
**Reproducibility**: I don't have a copy of TOUGHREACT, so I have reached out to the authors of the software, given them the files generated by the tutorials, and am waiting to hear back from them on whether the results are valid inputs for TOUGHREACT

## Functionality
**Functionality**: Same as reproducibility

## Documentation
**Functionality docomentation**: A few of the submodules are not documented, but the documentation is in a much better state than at the beginning of the review
**Automated testing**: The tests that are included don't prove that the software is functional yet

## Software paper
This needs a lot of work, and I created an issue tracking the acceptability criteria here: https://github.com/temmy222/PyTOUGHREACT/issues/18

After these issues are resolved, I will be ready to conclude my review.
