### 2024-05-02 temmy222/PyTOUGHREACT#8 @frank1010111
Can we reopen this issue? @matthewfeickert makes my argument more eloquently than I could.
