+++ src/pytoughreact/writers/react_writing.py
@@ -35,6 +35,7 @@ from pytoughreact.exceptions.custom_error import NotFoundError
 from pytoughreact.wrapper.reactzone import T2Zone
 from copy import deepcopy
 from t2data import t2data, fix_blockname, rocktype
+from t2grids import t2connection
 import numpy as np
 import os
 import sys
@@ -49,9 +50,18 @@ import typing
 
 class T2ReactParser(fixed_format_file):
     """Class for parsing REACTION data file."""
-    def __init__(self, filename, mode, read_function=default_read_function):
-        super(T2ReactParser, self).__init__(filename, mode,
-                                            t2react_format_specification, read_function)
+    def __init__(self, filename, mode, specification=t2react_format_specification, read_function=default_read_function, location=None):
+        # super(T2ReactParser, self).__init__(filename, mode,
+        #                                     t2react_format_specification, read_function)
+        self.specification = specification
+        self.read_function = read_function
+        self.preprocess_specification()
+        print(os.getcwd())
+        if location is None:
+            self.file = open(filename, mode)
+        else:
+            os.chdir(location)
+            self.file = open(filename, mode)
 
 
 class T2ExtraPrecisionDataParser(fixed_format_file):
@@ -110,10 +120,28 @@ class T2React(t2data):
         self._extra_precision, self._echo_extra_precision = [], True
         self.update_read_write_functions()
         self.read_function = read_function
-        if self.filename:
-            self.read(filename, meshfilename)
+        # if self.filename:
+        #     self.read(filename, meshfilename)
         self.minerals = []
         self.all_species = []
+
+    def read_connections(self, infile):
+        """Reads grid connections"""
+        self.grid.connectionlist, self.grid.connection = [], {}
+        line = padstring(infile.readline())
+        while line.strip() and not line.startswith('+++'):
+            [name1, name2, nseq,
+             nad1, nad2, isot, d1, d2,
+             areax, betax, sigx] = infile.parse_string(line, 'connections')
+            name1, name2 = fix_blockname(name1), fix_blockname(name2)
+            if nseq == 0: nseq = None
+            if nad1 == 0: nad1 = None
+            if nad2 == 0: nad2 = None
+            self.grid.add_connection(t2connection([self.grid.block[name1],
+                                                   self.grid.block[name2]],
+                                                  isot, [d1, d2], areax, betax,
+                                                  sigx, nseq, nad1, nad2))
+            line = padstring(infile.readline())
 
     def read_rocktypes(self, infile):
         """ Reads grid rock types
