+++ tox.ini
@@ -28,6 +28,7 @@ deps =
     PyTOUGH
     flake8
     pep8-naming
+    pytest-mock
 allowlist_externals=*
 commands =
     pytest --verbose
