+++ test/test_pytoughreact_react.py
@@ -343,7 +343,8 @@ class ReactTestCase():
 
     def set_up_read(self):
         react = T2React()
-        react.read('flow2.inp')
+        file_path = os.path.dirname(os.path.realpath(__file__))
+        react.read('flow2.inp', file_location=file_path)
         write_chemical = T2Chemical(t2reactgrid=react.grid)
         write_solute = T2Solute(t2chemical=write_chemical)
 
