### 2024-05-11 f9360bc
updates tests section in documentation
