+++ documentation/developer.rst
@@ -35,6 +35,7 @@ Updates to developer documentation are automatically uploaded to https://pytough
 Test
 ------------------------------
 Testing for the software was done using pytest. Before test can be run, pytest has to be
-installed using `pip install pytest`. The tests can be run by navigating to
-the root folder and running `pytest`. The results are then shown together with the 
+installed using `pip install pytest`. Coverage for the test is also analyzed using the 
+`pytest-cov` library.  This library needs to be installed using `pip install pytest-cov`.The tests can be run by navigating to
+the root folder and running `pytest` The results are then shown together with the 
 coverage of the tests. The settings for the tests can be modified through pyproject.toml
