+++ documentation/contributing.rst
@@ -1,5 +1,10 @@
-Contributions
 ==============
+Community Guidelines
+==============
+
+
+Contributing
+=============
 
 The project code is domiciled at the GitHub repo https://github.com/temmy222/PyTOUGHREACT 
 
@@ -7,13 +12,16 @@ Pull requests are welcome. For major changes, please open an issue first to disc
 
 Contributing to the code would involve you following the below procedures to quickly get started
 
-1. Clone the repo using preferred cloning method
+1. Clone the repo using preferred cloning method.
+
 2. Install the library to enable you able to use the test example using
 
 ```python
 pip install -e .
 ```
-3. Modify the code 
+
+3. Modify the code
+
 4. Tests are conducted with pytest and coverage reports are performed using pytest-cov. Install pytest and pytest-cov using the commands below
    
 ```python
@@ -23,6 +31,7 @@ pip install pytest
 ```python
 pip install pytest-cov
 ```
+
 5. Run tests:  Run the below command from the root folder to run the tests
    
 ```python
@@ -47,4 +56,5 @@ pip install pep8-naming
 ```
 
 8. Make a pull request after passing all tests
+
 9. More information can be found in developer notes in the documentation - https://pytoughreact.readthedocs.io/en/master/developer.html 
