### 2024-05-11 temmy222/PyTOUGHREACT#20 @temmy222
Hi @blakeaw . I added community guidelines to both the documentation and the repo. The documentation has also been changed to the master sub domain here - https://pytoughreact.readthedocs.io/en/master/contributing.html
