### 2024-05-11 temmy222/PyTOUGHREACT#19 @temmy222
Hi @blakeaw . Requirements for pytest-cov have been added to the documentation
