### 2024-05-11 eee975f
update technical.rst
