### 2024-05-11 1cbc0ac
update technical.rst
