### 2024-05-11 4e84d06
update technical.rst
