### 2024-05-11 7bc0f7b
update technical.rst
