+++ docs/publication/paper.md
@@ -28,29 +28,29 @@ The suite of TOUGH simulators by the Lawrence Berkeley National Laboratory (LBNL
 
 TOUGH suite of simulators requires users to write or modify a text file and parse the file to an executable to perform the simulation. Results from simulations are subsequently also saved to text files for interpretation by third party softwares. These makes it cumbersome for sensitivity analysis and uncertainty studies to be conducted as this would require a user to make multiple manual edits to files to extract required data. As would be imagined, this is also prone to human errors. Further, coupling the simulator with other simulators would be more involving. As a result of this, numerous tools have been created that wrap around the executables and make writing, reading and visualization of model properties easier to implement. Examples of such include PetraSim [@Yamamoto:2008] for which assist in creating the models and visualizing results, TOUGH2VIEWER [@Bondua:2012] and TECPLOT [@tecplot] for postprocessing and visualizations and IGMESH for building and integrating visualization tools suited especially for irregular gridding [@Hu:2016]. For scripting purposes, PyTOUGH [@Croucher:2011] and TOUGHIO [@Luu:2020] are used for pre and post processing of TOUGH2 simulations. So far, no tool exists for scripting of the reaction softwares of TOUGH. Reactive processes as with other processes are subject to lots of uncertainties and need to be adequately accounted for in engineering studies. To enable this with the TOUGHREACT simulator, it is essential to create a scripting tool to accomplish this. This tool extends the work done by PyTOUGH in creating an automatic platform for running TOUGH flow simulations by creating a concurrent tool for automating chemical and biodegradation reactions from any python enabled terminal or IDE.  
 
-# Architecture
+<!-- # Architecture
 
 ## General Architecture
 
-The software uses object-oriented programming principles to structure the code. The software can be thought of as composed of two main segments; the processing segment and the output segment (\autoref{fig:Figure 1}). The processing segment contains three main sections, IO processing, BIO and REACT.  The IO processing is responsible for most of the input and output processing such as reading, writing. The BIO section is responsible for the TMVOC section of the package where it contains classes for storing biomass and degradation information and processing it before passing to IO processing for read/write. Similarly, the react section assists in processing reaction parameters such as mineral, chemical, and solute information before passing to the IO processing segment for read/write. After the files have been written to or read from the appropriate file types, the executable is called from within PyTOUGHREACT and the simulation is performed using the executable. Thereafter, the output segment is called which can read the results of the simulations and contains methods and functions which assist the user in creating 2D or 3D plots through the plotting module.
+The software uses object-oriented programming principles to structure the code. The software can be thought of as composed of two main segments; the processing segment and the output segment (\autoref{fig:Figure 1}). The processing segment contains three main sections, IO processing, BIO and REACT.  The IO processing is responsible for most of the input and output processing such as reading, writing. The BIO section is responsible for the TMVOC section of the package where it contains classes for storing biomass and degradation information and processing it before passing to IO processing for read/write. Similarly, the react section assists in processing reaction parameters such as mineral, chemical, and solute information before passing to the IO processing segment for read/write. After the files have been written to or read from the appropriate file types, the executable is called from within PyTOUGHREACT and the simulation is performed using the executable. Thereafter, the output segment is called which can read the results of the simulations and contains methods and functions which assist the user in creating 2D or 3D plots through the plotting module. -->
 
-![General Structure of the PyTOUGHREACT package.\label{fig:Figure 1}](../images/general_architecture.png)
+<!-- ![General Structure of the PyTOUGHREACT package.\label{fig:Figure 1}](../images/general_architecture.png)
 
 ## BIO Architecture
 
[…]
