+++ docs/publication/paper.bib
@@ -44,3 +44,12 @@
   year = {2013},
   publisher = {Tecplot},
 }
+
+@article{Ajayi:2021, 
+  	Author = {{Ajayi}, T. and {Gupta},I.},
+  	Journal = {Journal of Petroleum Science and Engineering},
+  	Keywords = {Reactive Transport Modeling},
+  	Title = {{Long term assessment of the geochemical integrity of offshore wellbore cement– results from numerical modeling}},
+  	doi = {10.1016/j.petrol.2021.108443},
+  	Year = 2021
+}
