+++ docs/publication/paper.md
@@ -26,7 +26,16 @@ The suite of TOUGH simulators by the Lawrence Berkeley National Laboratory (LBNL
 
 # Introduction
 
-TOUGH suite of simulators requires users to write or modify a text file and parse the file to an executable to perform the simulation. Results from simulations are subsequently also saved to text files for interpretation by third party softwares. These makes it cumbersome for sensitivity analysis and uncertainty studies to be conducted as this would require a user to make multiple manual edits to files to extract required data. As would be imagined, this is also prone to human errors. Further, coupling the simulator with other simulators would be more involving. As a result of this, numerous tools have been created that wrap around the executables and make writing, reading and visualization of model properties easier to implement. Examples of such include PetraSim [@Yamamoto:2008] for which assist in creating the models and visualizing results, TOUGH2VIEWER [@Bondua:2012] and TECPLOT [@tecplot] for postprocessing and visualizations and IGMESH for building and integrating visualization tools suited especially for irregular gridding [@Hu:2016]. For scripting purposes, PyTOUGH [@Croucher:2011] and TOUGHIO [@Luu:2020] are used for pre and post processing of TOUGH2 simulations. So far, no tool exists for scripting of the reaction softwares of TOUGH. Reactive processes as with other processes are subject to lots of uncertainties and need to be adequately accounted for in engineering studies. To enable this with the TOUGHREACT simulator, it is essential to create a scripting tool to accomplish this. This tool extends the work done by PyTOUGH in creating an automatic platform for running TOUGH flow simulations by creating a concurrent tool for automating chemical and biodegradation reactions from any python enabled terminal or IDE.  
+TOUGH suite of simulators requires users to write or modify a text file and parse the file to an executable to perform the simulation. Results from simulations are subsequently also saved to text files for interpretation by third party softwares. These make it cumbersome for sensitivity analysis and uncertainty studies to be conducted as this would require a user to make multiple manual edits to files to extract required data. As would be imagined, this is also prone to human errors. Further, coupling the simulator with other simulators would be more involving. PyTOUGHREACT solves this problem by taking 
+away from the user the manual task of writing to file, reading from file and
+typing out command line statements to run the simulator. It 
+also assists users in processing the results from the simulation runs. It does so
+by providing a set of Python classes and functions in which users can use to 
+access file contents quickly. Since it uses Python to quickly perform these tasks, it can also be used with sensitivity and uncertainty quantification libraries to assess sensitivities and uncertainties in simulations without manually editing each file.
+
[…]
