### 2024-05-11 temmy222/PyTOUGHREACT#18 @temmy222
Hi @frank1010111 . I have made adjustments to the paper

1) I have moved the architecture and examples out of the paper and into the documentations or in the repo
2) The introduction has been broken down to three sections to reflect the points you highlighted.

Thank you
