### 2024-05-11 #6486 @temmy222
Hi @frank1010111

I added additional testing to test different pieces of the software independently. I will continue to add more tests. The coverage for the tests is now about 70%.

The documentation has been further improved.

The paper has been modified based on the checklist you specified in the issue.
