### 2024-05-15 temmy222/PyTOUGHREACT#20 @blakeaw
@temmy222, cool. However, I guess the Documentation part of the readme needs to be updated now:

> Documentation
>
> Documentation can be found here https://pytoughreact.readthedocs.io/en/latest/

The original readthedocs link now returns a 404 error.
