### 2024-05-15 temmy222/PyTOUGHREACT#22 @blakeaw
JOSS Review: API Documentation for Class definitions
This Issue has been raised as a part of the PyTOUGHREACT JOSS Review at: https://github.com/openjournals/joss-reviews/issues/6486

I think the API documentation could use a little more detail, particularly for all the classes.

One issue is that many class definitions are still missing the class-level docstrings. For example, the `Water` class has no class-level docstring.
https://pytoughreact.readthedocs.io/en/master/chemical.html#chemical-composition

Another issue is that some of the ones with class-level docstrings have descriptions that are too short to be helpful for new users/developers. For example, the `BIODG` and `Process` classes both just say
> Process specification
in the class-level docstring.
https://pytoughreact.readthedocs.io/en/master/chemical.html#biodegradation-processes
In this case, the docstring should give some additional details so it's clearer from the API documentation how these two classes differ.
