### 2024-05-21 #6486 @matthewfeickert
@frank1010111 @blakeaw Thank you for your continued reviews, and thank you all for your patience with my lack of responsiveness (I have been traveling for work internationally for the last weeks and my email is unfortunately quite a mess at the moment).

> **Reproducibility**: I don't have a copy of TOUGHREACT, so I have reached out to the authors of the software, given them the files generated by the tutorials, and am waiting to hear back from them on whether the results are valid inputs for TOUGHREACT

@frank1010111 Thank you for raising this point. Until now it has not been clear to me that [LBNL's `TOUGHREACT` is not open source](https://tough.lbl.gov/licensing-download/toughreact-licensing-download/). This information is missing from the paper draft and should be disclosed. I will consult with the other JOSS editors on how to proceed.

---

@temmy222 I would recommend that you make PRs that target the Issues that the reviewers open up (like https://github.com/temmy222/PyTOUGHREACT/issues/18) and ask them for clarification or review in their Issues before you close them. This will make the reviewer work much easier and more efficient.
