### 2024-05-22 #6486 @matthewfeickert
@temmy222 The dependency on `TOUGHREACT` is currently a problem, as it is [closed source and proprietary](https://tough.lbl.gov/licensing-download/toughreact-licensing-download/) and the reviewers are required to be able to test the functionality of the software.

As noted in the [Submitting a paper to JOSS docs](https://joss.readthedocs.io/en/latest/submitting.html#submissions-using-proprietary-languages-development-environments)

> ### Submissions using proprietary languages/development environments
>
> We strongly prefer software that doesn’t rely upon proprietary (paid for) development environments/programming languages. However, provided _your submission meets our requirements_ (including having a valid open source license) then we will consider your submission for review. Should your submission be accepted for review, we may ask you, the submitting author, to help us find reviewers who already have the required development environment installed.

All of the reviewers are going to need to be able to have at least runtime access to `TOUGHREACT` for the review to proceed. Please let us know what your plan is here. I would suggest starting by contacting LBL RE: seeing if there is any special consideration for licensing for academic journal reviews.

Please also reopen https://github.com/temmy222/PyTOUGHREACT/issues/8 as this can not be closed until this is resolved.
