### 2024-06-01 temmy222/PyTOUGHREACT#8 @temmy222
@matthewfeickert  @frank1010111. Additional tests have been added to the software which have improved the code coverage
