### 2024-06-02 temmy222/PyTOUGHREACT#8 @matthewfeickert
@temmy222 Please reread https://github.com/temmy222/PyTOUGHREACT/issues/8#issuecomment-2087739837 and then reopen this Issue.

Please also keep my comments in https://github.com/openjournals/joss-reviews/issues/6486#issuecomment-2121848175 in mind when interacting with the reviewers.
