### 2024-08-29 #6486 @temmy222
Hi @matthewfeickert. I still havent heard back from the authors of TOUGHREACT. Since this is a requirement for publishing in JOSS, I will like to withdraw the paper for publication. Thank you to everyone who have spent time reviewing the paper. The quality of the software has improved as a result of the reviews
