### 2024-09-03 #6486 @matthewfeickert
@openjournals/pe-eics Can you please withdraw the paper given https://github.com/openjournals/joss-reviews/issues/6486#issuecomment-2316623003?
