### 2024-09-03 #6486 @kyleniemeyer
@temmy222 although publication in JOSS isn't possible for your package at this time, you can still make your software citeable by following [GitHub's guide to creating a permanent archive via Zenodo](https://guides.github.com/activities/citable-code/).
